### 2025-05-06 f0d1100
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 7741b694a60ce4f68ce43b897684c16263d06f6a
Branch  : main
Author  : Matt Graham <matthew.m.graham@gmail.com>
Time    : 2025-05-06 05:34:53 +0000
Message : Update link to repository README (#570)
