+++ md5sum.txt
@@ -2,7 +2,7 @@
 "CODE_OF_CONDUCT.md" "c93c83c630db2fe2462240bf72552548" "site/built/CODE_OF_CONDUCT.md" "2025-02-11"
 "LICENSE.md" "b24ebbb41b14ca25cf6b8216dda83e5f" "site/built/LICENSE.md" "2025-02-11"
 "config.yaml" "deddc9218bcf80fc36f3f391f54f2675" "site/built/config.yaml" "2025-02-11"
-"index.md" "b326559d728c59d2e77a1db0b8a63ab9" "site/built/index.md" "2025-02-11"
+"index.md" "8b5609014b8028029f48266bc751663e" "site/built/index.md" "2025-05-06"
 "links.md" "8184cf4149eafbf03ce8da8ff0778c14" "site/built/links.md" "2025-02-11"
 "workshops.md" "912f39df323e22bb14340184cdd139f6" "site/built/workshops.md" "2025-02-11"
 "episodes/1-introduction.md" "8dabfa4853b660c8bfcb0aea5f435029" "site/built/1-introduction.md" "2025-02-11"
