### 2025-05-06 805b477
site deploy
Auto-generated via `{sandpaper}`
Source  : f0d110006a8fd6fe127db0419e55d5548a9db71e
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-05-06 05:35:38 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : 7741b694a60ce4f68ce43b897684c16263d06f6a
Branch  : main
Author  : Matt Graham <matthew.m.graham@gmail.com>
Time    : 2025-05-06 05:34:53 +0000
Message : Update link to repository README (#570)
