### 2025-05-07 d24e048
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 56707d4daa5cf9d56ffc876343292bc5c303cba4
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-05-07 14:32:59 +0000
Message : Merge pull request #580 from carpentries-lab/code4yonglei-patch-1

Update workshops.md
