+++ md5sum.txt
@@ -4,7 +4,7 @@
 "config.yaml" "deddc9218bcf80fc36f3f391f54f2675" "site/built/config.yaml" "2025-02-11"
 "index.md" "8b5609014b8028029f48266bc751663e" "site/built/index.md" "2025-05-06"
 "links.md" "8184cf4149eafbf03ce8da8ff0778c14" "site/built/links.md" "2025-02-11"
-"workshops.md" "912f39df323e22bb14340184cdd139f6" "site/built/workshops.md" "2025-02-11"
+"workshops.md" "ba594ced8acc714eb6785d1c63191bb8" "site/built/workshops.md" "2025-05-07"
 "episodes/1-introduction.md" "8dabfa4853b660c8bfcb0aea5f435029" "site/built/1-introduction.md" "2025-02-11"
 "episodes/2-keras.md" "ebddf0ec35a0e97735a1ea7f1d204a40" "site/built/2-keras.md" "2025-02-11"
 "episodes/3-monitor-the-model.md" "93984f2bd862ddc2f10ba6749950b719" "site/built/3-monitor-the-model.md" "2025-02-11"
