### 2025-05-13 c32ab88
site deploy
Auto-generated via `{sandpaper}`
Source  : d24e04855eb35485b7ab0e617a3be867a10d14b0
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-05-07 14:33:56 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : 56707d4daa5cf9d56ffc876343292bc5c303cba4
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-05-07 14:32:59 +0000
Message : Merge pull request #580 from carpentries-lab/code4yonglei-patch-1

Update workshops.md
