### 2025-05-13 b821a55
Update setup.md
1. Create folder for workshop - store data and venv folder
2. Call virtual environment folder "venv", following conventions
3. Require Python 3.11.9. Newer versions of Python are not yet compatible with Tensorflow
