### 2025-05-13 2ea33dd
Update setup.md
