### 2025-05-13 0386df9
add NLP workshop
