### 2025-05-13 287e0be
fix headers/link formatting to match pre-existing examples
