### 2025-05-14 d978106
move prereqs before other related lessons
