### 2025-05-14 0c0639c
fix bulleted list formatting
