+++ learners/setup.md
@@ -163,7 +163,7 @@ An [optional challenge in episode 2](episodes/2-keras.md) requires installation 
 and instructions for doing that can be found
 [by following this link](https://graphviz.org/download/).
 
-## Starting Jupyter Lab
+### Starting Jupyter Lab
 
 We will teach using Python in [Jupyter Lab][jupyter], a programming environment that runs in a web browser.
 Jupyter Lab is compatible with Firefox, Chrome, Safari and Chromium-based browsers.
@@ -178,7 +178,7 @@ and type the command:
 jupyter lab
 ```
 
-## Check your setup
+### Check your virtual software setup
 To check whether all packages installed correctly, start a jupyter notebook in jupyter lab as
 explained above (**with virtual environment activated**). Run the following check to verify you have the right version of Python configured.
 ```python
@@ -211,7 +211,7 @@ Most versions will work fine with this lesson, but:
 - For Keras and Tensorflow, the minimum version is 2.12.0
 - For sklearn, the minimum version is 1.2.2
 
-## Fallback option: cloud environment
+### Fallback option: cloud environment
 If a local installation does not work for you, it is also possible to run this lesson in [Binder Hub](https://mybinder.org/v2/gh/carpentries-incubator/deep-learning-intro/scaffolds). This should give you an environment with all the required software and data to run this lesson, nothing which is saved will be stored, please copy any files you want to keep. Note that if you are the first person to launch this in the last few days it can take several minutes to startup. The second person who loads it should find it loads in under a minute. Instructors who intend to use this option should start it themselves shortly before the workshop begins.
 
 Alternatively you can use [Google colab](https://colab.research.google.com/). If you open a jupyter notebook here, the required packages are already pre-installed. Note that google colab uses jupyter notebook instead of Jupyter Lab.
