+++ md5sum.txt
@@ -7,7 +7,7 @@
 "workshops.md" "ba594ced8acc714eb6785d1c63191bb8" "site/built/workshops.md" "2025-05-07"
 "episodes/1-introduction.md" "8dabfa4853b660c8bfcb0aea5f435029" "site/built/1-introduction.md" "2025-02-11"
 "episodes/2-keras.md" "ebddf0ec35a0e97735a1ea7f1d204a40" "site/built/2-keras.md" "2025-02-11"
-"episodes/3-monitor-the-model.md" "93984f2bd862ddc2f10ba6749950b719" "site/built/3-monitor-the-model.md" "2025-02-11"
+"episodes/3-monitor-the-model.md" "6e3ad68bafda09d90a9682dd576795dc" "site/built/3-monitor-the-model.md" "2025-05-15"
 "episodes/4-advanced-layer-types.md" "97b49e9dad76479bcfe608f0de2d52a4" "site/built/4-advanced-layer-types.md" "2025-02-11"
 "episodes/5-transfer-learning.md" "5808f2218c3f2d2d400e1ec1ad9f1f3c" "site/built/5-transfer-learning.md" "2025-02-11"
 "episodes/6-outlook.md" "007728216562f3b52b983ff1908af5b7" "site/built/6-outlook.md" "2025-02-11"
