+++ md5sum.txt
@@ -9,7 +9,7 @@
 "episodes/2-keras.md" "ebddf0ec35a0e97735a1ea7f1d204a40" "site/built/2-keras.md" "2025-02-11"
 "episodes/3-monitor-the-model.md" "93984f2bd862ddc2f10ba6749950b719" "site/built/3-monitor-the-model.md" "2025-02-11"
 "episodes/4-advanced-layer-types.md" "97b49e9dad76479bcfe608f0de2d52a4" "site/built/4-advanced-layer-types.md" "2025-02-11"
-"episodes/5-transfer-learning.md" "5808f2218c3f2d2d400e1ec1ad9f1f3c" "site/built/5-transfer-learning.md" "2025-02-11"
+"episodes/5-transfer-learning.md" "6118c2c2d1c86ec045309932b6edba8c" "site/built/5-transfer-learning.md" "2025-05-15"
 "episodes/6-outlook.md" "007728216562f3b52b983ff1908af5b7" "site/built/6-outlook.md" "2025-02-11"
 "instructors/bonus-material.md" "382832ea4eb097fc7781cb36992c1955" "site/built/bonus-material.md" "2025-02-11"
 "instructors/design.md" "1537f9d0c90cdfdb1781bd3daf94dadd" "site/built/design.md" "2025-02-11"
