+++ md5sum.txt
@@ -1,0 +1,22 @@
+"file" "checksum" "built" "date"
+"CODE_OF_CONDUCT.md" "c93c83c630db2fe2462240bf72552548" "site/built/CODE_OF_CONDUCT.md" "2025-02-11"
+"LICENSE.md" "b24ebbb41b14ca25cf6b8216dda83e5f" "site/built/LICENSE.md" "2025-02-11"
+"config.yaml" "deddc9218bcf80fc36f3f391f54f2675" "site/built/config.yaml" "2025-02-11"
+"index.md" "8b5609014b8028029f48266bc751663e" "site/built/index.md" "2025-05-06"
+"links.md" "8184cf4149eafbf03ce8da8ff0778c14" "site/built/links.md" "2025-02-11"
+"workshops.md" "ba594ced8acc714eb6785d1c63191bb8" "site/built/workshops.md" "2025-05-07"
+"episodes/1-introduction.md" "8dabfa4853b660c8bfcb0aea5f435029" "site/built/1-introduction.md" "2025-02-11"
+"episodes/2-keras.md" "ebddf0ec35a0e97735a1ea7f1d204a40" "site/built/2-keras.md" "2025-02-11"
+"episodes/3-monitor-the-model.md" "93984f2bd862ddc2f10ba6749950b719" "site/built/3-monitor-the-model.md" "2025-02-11"
+"episodes/4-advanced-layer-types.md" "97b49e9dad76479bcfe608f0de2d52a4" "site/built/4-advanced-layer-types.md" "2025-02-11"
+"episodes/5-transfer-learning.md" "5808f2218c3f2d2d400e1ec1ad9f1f3c" "site/built/5-transfer-learning.md" "2025-02-11"
+"episodes/6-outlook.md" "007728216562f3b52b983ff1908af5b7" "site/built/6-outlook.md" "2025-02-11"
+"instructors/bonus-material.md" "382832ea4eb097fc7781cb36992c1955" "site/built/bonus-material.md" "2025-02-11"
+"instructors/design.md" "1537f9d0c90cdfdb1781bd3daf94dadd" "site/built/design.md" "2025-02-11"
+"instructors/instructor-notes.md" "73b56a12ce196586a580b9399e2d49fc" "site/built/instructor-notes.md" "2025-02-11"
+"instructors/schedule.md" "332b32d24f144b29a280176e6b5d015f" "site/built/schedule.md" "2025-03-10"
+"instructors/survey-templates.md" "ea5d46e7b54d335f79e57a7bc31d1c5c" "site/built/survey-templates.md" "2025-02-11"
+"learners/reference.md" "e47218673643f23431c540c2d3b27868" "site/built/reference.md" "2025-02-11"
+"learners/setup.md" "f640f3d679d73aad199d11e53590cb3a" "site/built/setup.md" "2025-02-11"
+"paper/paper.md" "5257d2bb9c07be835c20d66855c38202" "site/built/paper.md" "2025-03-03"
+"profiles/learner-profiles.md" "ef0f26dd0874387d80ed3fd468b99e23" "site/built/learner-profiles.md" "2025-02-11"
