+++ index.md
@@ -1,0 +1,60 @@
+---
+site: sandpaper::sandpaper_site
+---
+## Welcome
+This is a hands-on introduction to the first steps in deep learning, intended for researchers who are familiar with (non-deep) machine learning.
+
+The use of deep learning has seen a sharp increase of popularity and applicability over the last decade. 
+While deep learning can be a useful tool for researchers from a wide range of domains, 
+taking the first steps in the world of deep learning can be somewhat intimidating. 
+This introduction covers the basics of deep learning in a practical and hands-on manner, 
+so that upon completion, you will be able to train your first neural network and understand what next steps to take to improve the model.
+
+We start with explaining the basic concepts of neural networks, and then go through the different steps of a deep learning workflow. 
+Learners will learn how to prepare data for deep learning, how to implement a basic deep learning model in Python with Keras, 
+how to monitor and troubleshoot the training process and how to implement different layer types such as convolutional layers.
+
+### Other related lessons
+#### Introduction to artificial neural networks in Python
+The [Introduction to artificial neural networks in Python lesson](https://carpentries-incubator.github.io/machine-learning-neural-python/)
+takes a different angle to introducing deep learning, 
+focusing on computer vision with the application on medical images.
+
+#### Introduction to machine learning in Python with scikit-learn
+The [Introduction to machine learning in Python with scikit-learn lesson](https://esciencecenter-digital-skills.github.io/scikit-learn-mooc/)
+introduces practical machine learning using Python. It is a good lesson to follow in preparation for this lesson,
+since basic knowledge of machine learning and Python programming skills are required for this lesson.
+
+:::::::::::::::::: checklist
+
+## Prerequisites
+Learners are expected to have the following knowledge:
+
+- Basic Python programming skills and familiarity with the Pandas package.
+- Basic knowledge on machine learning, including the following concepts: Data cleaning, train & test split, type of problems (regression, classification), overfitting & underfitting, metrics (accuracy, recall, etc.).
+
+::::::::::::::::::::::::::::
+
+::: instructor
+
+## We can help you out with teaching this lesson
+
+Do you want to teach this lesson?
+Find more help in the [README](https://github.com/carpentries-lab/deep-learning-intro?tab=readme-ov-file#teaching-this-lesson)
+Feel free to reach out to us with any questions that you have.
+Just open a new issue.
+We also value any feedback on the lesson!
+
+:::
+
+::: instructor
+
+## Breaks and Schedule
+
+Episode 2, 3, and 4 in this lesson are relatively long.
+We suggest to have a break at least every 90 minutes and to switch the instructor regularly, also within episodes.
+We have added reminders to the longer episodes with suggestions for when to have a switch and/or a break.
+
+There is an [example schedule](schedule.html) with breaks that can be adapted to how you want to teach the lesson.
+
+:::
