+++ 2-keras.md
@@ -1,0 +1,904 @@
+---
+title: "Classification by a neural network using Keras"
+teaching: 60
+exercises: 50
+---
+
+::: questions
+- How do I compose a neural network using Keras?
+- How do I train this network on a dataset?
+- How do I get insight into learning process?
+- How do I measure the performance of the network?
+:::
+
+::: objectives
+- Use the deep learning workflow to structure the notebook
+- Explore the dataset using pandas and seaborn
+- Identify the inputs and outputs of a deep neural network.
+- Use one-hot encoding to prepare data for classification in Keras
+- Describe a fully connected layer
+- Implement a fully connected layer with Keras
+- Use Keras to train a small fully connected network on prepared data
+- Interpret the loss curve of the training process
+- Use a confusion matrix to measure the trained networks' performance on a test set
+:::
+
+
+## Introduction
+In this episode we will learn how to create and train a neural network using Keras to solve a simple classification task.
+
+The goal of this episode is to quickly get your hands dirty in actually defining and training a neural network,
+without going into depth of how neural networks work on a technical or mathematical level.
+We want you to go through the full deep learning workflow once before going into more details.
+
+In fact, this is also what we would recommend you to do when working on real-world problems:
+First quickly build a working pipeline, while taking shortcuts.
+Then, slowly make the pipeline more advanced while you keep on evaluating the approach.
+
+In [episode 3](episodes/3-monitor-the-model.md) we will expand on the concepts that are lightly introduced in this episode.
+Some of these concepts include: how to monitor the training progress and how optimization works.
+
+::: instructor
+It is good to stress the goal for this episode a few times, because learners will usually have a lot of questions like:
+'Why don't we normalize our features' or 'Why do we choose Adam optimizer?'.
+It can be a good idea to park some of these questions for discussion in episode 3 and 4.
+:::
+
+As a reminder below are the steps of the deep learning workflow:
+
+1. Formulate / Outline the problem
+2. Identify inputs and outputs
+3. Prepare data
+4. Choose a pretrained model or start building architecture from scratch
+5. Choose a loss function and optimizer
+6. Train the model
+7. Perform a Prediction/Classification
+8. Measure performance
+9. Refine the model
+10. Save model
+
+In this episode we will focus on a minimal example for each of these steps, later episodes will build on this knowledge to go into greater depth for some or all of these steps.
+
+::: callout
+## GPU usage
+For this lesson having a [GPU (graphics processing unit)](https://glosario.carpentries.org/en/#gpu) available is not needed.
+We specifically use very small toy problems so that you do not need one.
+However, Keras will use your GPU automatically when it is available.
+Using a GPU becomes necessary when tackling larger datasets or complex problems which
+require a more complex neural network.
+:::
+
+## 1. Formulate/outline the problem: penguin classification
+In this episode we will be using the [penguin dataset](https://zenodo.org/record/3960218). This is a dataset that was published in 2020 by Allison Horst and contains data on three different species of the penguins.
+
+We will use the penguin dataset to train a neural network which can classify which species a
+penguin belongs to, based on their physical characteristics.
+
+::: callout
+## Goal
+The goal is to predict a penguins' species using the attributes available in this dataset.
+:::
+
+The `palmerpenguins` data contains size measurements for three penguin species observed on three islands in the Palmer Archipelago, Antarctica.
+The physical attributes measured are flipper length, beak length, beak width, body mass, and sex.
+
+![*Artwork by @allison_horst*][palmer-penguins]
+
+
+![*Artwork by @allison_horst*][penguin-beaks]
+
+
[…]
