### 2025-05-18 285e2d9
windows should use terminal as well. never use windows command prompt
