### 2025-05-18 0bb44ab
add min version numbers as comments
