+++ 4-advanced-layer-types.md
@@ -1,0 +1,1097 @@
+---
+title: "Advanced layer types"
+teaching: 70
+exercises: 70
+---
+
+::: questions
+- Why do we need different types of layers?
+- What are good network designs for image data?
+- What is a convolutional layer?
+- How can we use different types of layers to prevent overfitting?
+- What is hyperparameter tuning?
+:::
+
+::: objectives
+- Understand why convolutional and pooling layers are useful for image data
+- Implement a convolutional neural network on an image dataset
+- Use a dropout layer to prevent overfitting
+- Be able to tune the hyperparameters of a Keras model
+:::
+
+
+## Different types of layers
+Networks are like onions: a typical neural network consists of many layers. In fact, the word *deep* in *deep learning*
+refers to the many layers that make the network deep.
+
+So far, we have seen one type of layer, namely the **fully connected**, or **dense** layer. This layer is called fully connected, because all input neurons are taken into account by each output neuron. The number of parameters that need to be learned by the network, is thus in the order of magnitude of the number of input neurons times the number of hidden neurons.
+
+However, there are many different types of layers that perform different calculations and take different inputs. In this episode we will take a look at **convolutional layers** and **dropout layers**, which are useful in the context of image data, but also in many other types of (structured) data.
+
+## 1. Formulate / Outline the problem: Image classification
+The [MLCommons Dollar Street Dataset](https://www.kaggle.com/datasets/mlcommons/the-dollar-street-dataset) is a collection of images of everyday household items from homes around the world that visually captures socioeconomic diversity of traditionally underrepresented populations. We use [a subset of the original dataset](https://zenodo.org/records/10970014) that can be used for multiclass classification with 10 categories. Let's load the data:
+
+```python
+import pathlib
+import numpy as np
+
+DATA_FOLDER = pathlib.Path('data/dataset_dollarstreet/') # change to location where you stored the data
+train_images = np.load(DATA_FOLDER / 'train_images.npy')
+val_images = np.load(DATA_FOLDER / 'test_images.npy')
+train_labels = np.load(DATA_FOLDER / 'train_labels.npy')
+val_labels = np.load(DATA_FOLDER / 'test_labels.npy')
+```
+
+:::::::::::::::::::::::::::::::::::::::::: callout
+
+### A note about data provenance
+In an earlier version, this part of the lesson used a different example dataset.
+During [peer review](https://github.com/carpentries-lab/reviews/issues/25#issuecomment-1953271802), the decision was made to replace that dataset due to the way it had been compiled using images "scraped" from the internet without permission from or credit to the original creators of those images. Unfortunately, uncredited use of images is a common problem among datasets used to benchmark models for image classification.
+
+The Dollar Street dataset was chosen for use in the lesson as it contains only images [created by the Gapminder project](https://www.gapminder.org/dollar-street/about?) for the purposes of using them in the dataset.
+The original Dollar Street dataset is very large -- more than 100 GB -- with the potential to grow even bigger, so we created a subset for use in this lesson.
+
+::::::::::::::::::::::::::::::::::::::::::::::::::
+
+
+### Dollar Street 10 dataset
+
+The Dollar Street 10 dataset consists of images of 10 different classes, this is the mapping of the categories:
+
+| Category        | label |
+|-----------------|-------|
+| day bed         | 0     |
+| dishrag         | 1     |
+| plate           | 2     |
+| running shoe    | 3     |
+| soap dispenser  | 4     |
+| street sign     | 5     |
+| table lamp      | 6     |
[…]
