+++ setup.md
@@ -1,0 +1,154 @@
+---
+title: Setup
+---
+## Software Setup
+
+::::::::::::::::::::::::::::::::::::::: discussion
+
+### Installing Python
+
+[Python][python] is a popular language for scientific computing, and a frequent choice
+for machine learning as well.
+To install Python, follow the [Beginner's Guide](https://wiki.python.org/moin/BeginnersGuide/Download) or head straight to the [download page](https://www.python.org/downloads/).
+
+Please set up your python environment at least a day in advance of the workshop.
+If you encounter problems with the installation procedure, ask your workshop organizers via e-mail for assistance so
+you are ready to go as soon as the workshop begins.
+
+:::::::::::::::::::::::::::::::::::::::::::::::::::
+
+## Installing the required packages{#packages}
+
+[Pip](https://pip.pypa.io/en/stable/) is the package management system built into Python.
+Pip should be available in your system once you installed Python successfully.
+
+Open a terminal (Mac/Linux) or Command Prompt (Windows) and run the following commands.
+
+1. Create a [virtual environment](https://packaging.python.org/en/latest/guides/installing-using-pip-and-virtual-environments/#create-and-use-virtual-environments) called `dl_workshop`:
+
+::: spoiler
+
+### On Linux/macOs
+
+```shell
+python3 -m venv dl_workshop
+```
+
+:::
+
+::: spoiler
+
+### On Windows
+
+```shell
+py -m venv dl_workshop
+```
+
+:::
+
+2. Activate the newly created virtual environment:
+
+::: spoiler
+
+### On Linux/macOs
+
+```shell
+source dl_workshop/bin/activate
+```
+
+:::
+
+::: spoiler
+
+### On Windows
+
+```shell
+dl_workshop\Scripts\activate
+```
+
+:::
+
+Remember that you need to activate your environment every time you restart your terminal!
+
+3. Install the required packages:
+
+::: spoiler
+
+### On Linux/macOs
+
+```shell
+python3 -m pip install jupyter seaborn scikit-learn pandas tensorflow pydot
+```
+
+Note for MacOS users: there is a package `tensorflow-metal` which accelerates the training of machine learning models with TensorFlow on a recent Mac with a Silicon chip (M1/M2/M3).
+However, the installation is currently broken in the most recent version (as of January 2025), see the [developer forum](https://developer.apple.com/forums/thread/772147).
+
+:::
+
+::: spoiler
+
+### On Windows
+
+```shell
+py -m pip install jupyter seaborn scikit-learn pandas tensorflow pydot
+```
+
+:::
+
+Note: Tensorflow makes Keras available as a module too.
+
+An [optional challenge in episode 2](episodes/2-keras.md) requires installation of Graphviz
+and instructions for doing that can be found
+[by following this link](https://graphviz.org/download/).
+
+## Starting Jupyter Lab
+
+We will teach using Python in [Jupyter Lab][jupyter], a programming environment that runs in a web browser.
+Jupyter Lab is compatible with Firefox, Chrome, Safari and Chromium-based browsers.
+Note that Internet Explorer and Edge are *not* supported.
+See the [Jupyter Lab documentation](https://jupyterlab.readthedocs.io/en/latest/getting_started/accessibility.html#compatibility-with-browsers-and-assistive-technology) for an up-to-date list of supported browsers.
+
+To start Jupyter Lab, open a terminal (Mac/Linux) or Command Prompt (Windows), 
+make sure that you activated the virtual environment you created for this course,
+and type the command:
+
+```shell
+jupyter lab
+```
+
+## Check your setup
+To check whether all packages installed correctly, start a jupyter notebook in jupyter lab as
+explained above. Run the following lines of code:
+```python
+import sklearn
+print('sklearn version: ', sklearn.__version__)
+
+import seaborn
[…]
