+++ md5sum.txt
@@ -7,9 +7,9 @@
 "workshops.md" "ba594ced8acc714eb6785d1c63191bb8" "site/built/workshops.md" "2025-05-07"
 "episodes/1-introduction.md" "8dabfa4853b660c8bfcb0aea5f435029" "site/built/1-introduction.md" "2025-02-11"
 "episodes/2-keras.md" "ebddf0ec35a0e97735a1ea7f1d204a40" "site/built/2-keras.md" "2025-02-11"
-"episodes/3-monitor-the-model.md" "93984f2bd862ddc2f10ba6749950b719" "site/built/3-monitor-the-model.md" "2025-02-11"
-"episodes/4-advanced-layer-types.md" "97b49e9dad76479bcfe608f0de2d52a4" "site/built/4-advanced-layer-types.md" "2025-02-11"
-"episodes/5-transfer-learning.md" "5808f2218c3f2d2d400e1ec1ad9f1f3c" "site/built/5-transfer-learning.md" "2025-02-11"
+"episodes/3-monitor-the-model.md" "1763a6c4baeb9e1bcc08abf60a514d89" "site/built/3-monitor-the-model.md" "2025-05-18"
+"episodes/4-advanced-layer-types.md" "1f1911414502d92cc36eee139a751890" "site/built/4-advanced-layer-types.md" "2025-05-18"
+"episodes/5-transfer-learning.md" "7410e21a68de7ad807de4a543d970ded" "site/built/5-transfer-learning.md" "2025-05-18"
 "episodes/6-outlook.md" "007728216562f3b52b983ff1908af5b7" "site/built/6-outlook.md" "2025-02-11"
 "instructors/bonus-material.md" "382832ea4eb097fc7781cb36992c1955" "site/built/bonus-material.md" "2025-02-11"
 "instructors/design.md" "1537f9d0c90cdfdb1781bd3daf94dadd" "site/built/design.md" "2025-02-11"
@@ -17,6 +17,6 @@
 "instructors/schedule.md" "332b32d24f144b29a280176e6b5d015f" "site/built/schedule.md" "2025-03-10"
 "instructors/survey-templates.md" "ea5d46e7b54d335f79e57a7bc31d1c5c" "site/built/survey-templates.md" "2025-02-11"
 "learners/reference.md" "e47218673643f23431c540c2d3b27868" "site/built/reference.md" "2025-02-11"
-"learners/setup.md" "f640f3d679d73aad199d11e53590cb3a" "site/built/setup.md" "2025-02-11"
+"learners/setup.md" "759a539db318e12a287b57f724cdeb12" "site/built/setup.md" "2025-05-18"
 "paper/paper.md" "5257d2bb9c07be835c20d66855c38202" "site/built/paper.md" "2025-03-03"
 "profiles/learner-profiles.md" "ef0f26dd0874387d80ed3fd468b99e23" "site/built/learner-profiles.md" "2025-02-11"
