### 2025-05-19 0a49d41
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 72ca5c8e59ec3cc59e593cbe8197c0e0d8c27179
Branch  : main
Author  : Chris Endemann <endemann@wisc.edu>
Time    : 2025-05-19 18:25:21 +0000
Message : Merge pull request #589 from UW-Madison-DataScience/episode-4-updates

add missing import
