### 2025-07-18 c3052dd
Fixed some dead links (#590)
