+++ md5sum.txt
@@ -1,5 +1,5 @@
 "file" "checksum" "built" "date"
-"CODE_OF_CONDUCT.md" "c93c83c630db2fe2462240bf72552548" "site/built/CODE_OF_CONDUCT.md" "2025-02-11"
+"CODE_OF_CONDUCT.md" "dcaeeecd78bdc1bacd97d931c50e8269" "site/built/CODE_OF_CONDUCT.md" "2025-07-18"
 "LICENSE.md" "b24ebbb41b14ca25cf6b8216dda83e5f" "site/built/LICENSE.md" "2025-02-11"
 "config.yaml" "deddc9218bcf80fc36f3f391f54f2675" "site/built/config.yaml" "2025-02-11"
 "index.md" "8b5609014b8028029f48266bc751663e" "site/built/index.md" "2025-05-06"
@@ -13,7 +13,7 @@
 "episodes/6-outlook.md" "007728216562f3b52b983ff1908af5b7" "site/built/6-outlook.md" "2025-02-11"
 "instructors/bonus-material.md" "382832ea4eb097fc7781cb36992c1955" "site/built/bonus-material.md" "2025-02-11"
 "instructors/design.md" "1537f9d0c90cdfdb1781bd3daf94dadd" "site/built/design.md" "2025-02-11"
-"instructors/instructor-notes.md" "73b56a12ce196586a580b9399e2d49fc" "site/built/instructor-notes.md" "2025-02-11"
+"instructors/instructor-notes.md" "541a3ccba7b1edff5040ebfb5a1b0c3d" "site/built/instructor-notes.md" "2025-07-18"
 "instructors/schedule.md" "332b32d24f144b29a280176e6b5d015f" "site/built/schedule.md" "2025-03-10"
 "instructors/survey-templates.md" "ea5d46e7b54d335f79e57a7bc31d1c5c" "site/built/survey-templates.md" "2025-02-11"
 "learners/reference.md" "e47218673643f23431c540c2d3b27868" "site/built/reference.md" "2025-02-11"
