### 2025-07-24 #281 @labarba
@magsol – This lesson submission was pre-reviewed at the Carpentries Lab, and is submitted for publication in JOSE under our partnership with Carpentries. This means that we only assign one reviewer, who does a sanity check on the review history, and then ensures the JOSE publication criteria are met, but does not need to do a thorough review of the lesson materials. We build on the review conducted by the Carpentries, and our reviewer and editor ensure it is in-scope for JOSE, read the paper and check that it follows our author guidance (and has no formatting issues), check the archive and license are all in order, and recommends acceptance after any issues are resolved.
