### 2025-09-01 #281 @carschno
Hi @magsol,
Since the submission, a few things to the lesson have changed. Most importantly, the email of @svenvanderburg as specified above no longer works because Sven no longer works at the eScience Center. For the maintenance tasks, I am taking over as Sven mentioned [above](https://github.com/openjournals/jose-reviews/issues/281#issuecomment-2853374567). My email, if needed: c.schnober@esciencecenter.nl

Furthermore, we have two new co-authors that we would like to add to the authors list:
- Ashwin V. Mohanan (@ashwinvis)
- Yonglei Wang (@code4yonglei: could you confirm you want to be added too, please?)

@magsol regarding reviewers: Sven suggested a few candidates [above](https://github.com/openjournals/jose-reviews/issues/281#issuecomment-2853347095), not sure if they have been involved already. Is there anything we can help on that matter?
