### 2025-09-01 #281 @carschno
Closes https://github.com/carpentries-lab/deep-learning-intro/issues/568
