### 2025-09-01 1f4e276
markdown source builds
Auto-generated via `{sandpaper}`
Source  : a78f018aba5d4500d0e4fd2b719fc179fafaf662
Branch  : main
Author  : José Angel Daza <maldini.3@gmail.com>
Time    : 2025-09-01 11:43:12 +0000
Message : Merge pull request #592 from carpentries-lab/fix/links_#590

Fix/links #590
