### 2025-09-01 91b4171
site deploy
Auto-generated via `{sandpaper}`
Source  : 1f4e2764200196a824a6b4c05d71b415a978c1ff
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-01 11:44:12 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : a78f018aba5d4500d0e4fd2b719fc179fafaf662
Branch  : main
Author  : José Angel Daza <maldini.3@gmail.com>
Time    : 2025-09-01 11:43:12 +0000
Message : Merge pull request #592 from carpentries-lab/fix/links_#590

Fix/links #590
