### 2025-09-01 d4102ad
markdown source builds
Auto-generated via `{sandpaper}`
Source  : fd4afa03109f72e7ffcb49f2fc4d647de51496d7
Branch  : main
Author  : José Angel Daza <josdaza.a@gmail.com>
Time    : 2025-09-01 11:54:06 +0000
Message : Merge pull request #593 from carpentries-lab/548-set-random-seeds-in-episode-3-4-5

548 set random seeds in episode 3 4 5
