### 2025-09-01 12ac92d
site deploy
Auto-generated via `{sandpaper}`
Source  : d4102ade33fa5fc4b97d681a1951e3585fb44c53
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-01 11:54:52 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : fd4afa03109f72e7ffcb49f2fc4d647de51496d7
Branch  : main
Author  : José Angel Daza <josdaza.a@gmail.com>
Time    : 2025-09-01 11:54:06 +0000
Message : Merge pull request #593 from carpentries-lab/548-set-random-seeds-in-episode-3-4-5

548 set random seeds in episode 3 4 5
