+++ md5sum.txt
@@ -6,10 +6,10 @@
 "links.md" "8184cf4149eafbf03ce8da8ff0778c14" "site/built/links.md" "2025-02-11"
 "workshops.md" "ba594ced8acc714eb6785d1c63191bb8" "site/built/workshops.md" "2025-05-07"
 "episodes/1-introduction.md" "8dabfa4853b660c8bfcb0aea5f435029" "site/built/1-introduction.md" "2025-02-11"
-"episodes/2-keras.md" "ebddf0ec35a0e97735a1ea7f1d204a40" "site/built/2-keras.md" "2025-02-11"
-"episodes/3-monitor-the-model.md" "93984f2bd862ddc2f10ba6749950b719" "site/built/3-monitor-the-model.md" "2025-02-11"
-"episodes/4-advanced-layer-types.md" "5d0cfb4120210c34ae7d66c0ac3ddf81" "site/built/4-advanced-layer-types.md" "2025-05-19"
-"episodes/5-transfer-learning.md" "5808f2218c3f2d2d400e1ec1ad9f1f3c" "site/built/5-transfer-learning.md" "2025-02-11"
+"episodes/2-keras.md" "08518878cbf1344120181bf63b1a44f0" "site/built/2-keras.md" "2025-09-01"
+"episodes/3-monitor-the-model.md" "4be34c9bbba665390860faa71ccc22cc" "site/built/3-monitor-the-model.md" "2025-09-01"
+"episodes/4-advanced-layer-types.md" "61bb5fc50d1b2cde8163fdfff1a99bb0" "site/built/4-advanced-layer-types.md" "2025-09-01"
+"episodes/5-transfer-learning.md" "9dae01295f9f0a566a71c05672cc052f" "site/built/5-transfer-learning.md" "2025-09-01"
 "episodes/6-outlook.md" "007728216562f3b52b983ff1908af5b7" "site/built/6-outlook.md" "2025-02-11"
 "instructors/bonus-material.md" "382832ea4eb097fc7781cb36992c1955" "site/built/bonus-material.md" "2025-02-11"
 "instructors/design.md" "1537f9d0c90cdfdb1781bd3daf94dadd" "site/built/design.md" "2025-02-11"
