### 2025-09-01 182733b
Fix plot y axis label.
