### 2025-09-01 3aed2ef
expand alt text descriptions
