+++ episodes/3-monitor-the-model.md
@@ -46,7 +46,7 @@ Here we want to work with the *weather prediction dataset* (the light version) w
 It contains daily weather observations from 11 different European cities or places through the
 years 2000 to 2010. For all locations the data contains the variables ‘mean temperature’, ‘max temperature’, and ‘min temperature’. In addition, for multiple locations, the following variables are provided: 'cloud_cover', 'wind_speed', 'wind_gust', 'humidity', 'pressure', 'global_radiation', 'precipitation', 'sunshine', but not all of them are provided for every location. A more extensive description of the dataset including the different physical units is given in accompanying metadata file. The full dataset comprises of 10 years (3654 days) of collected weather data across Europe.
 
-![European locations in the weather prediction dataset](fig/03_weather_prediction_dataset_map.png){alt='18 European locations in the weather prediction dataset'}
+![European locations in the weather prediction dataset](fig/03_weather_prediction_dataset_map.png){alt='18 European locations in the weather prediction dataset distributed across Austria, France, Germany, Hungary, Italy, the Netherlands, Norway, Slovenia, Sweden, Switzerland, and the United Kingdom.'}
 
  A very common task with weather data is to make a prediction about the weather sometime in the future, say the next day. In this episode, we will try to predict tomorrow's sunshine hours, a challenging-to-predict feature, using a neural network with the available weather data for one location: BASEL.
 
@@ -261,14 +261,15 @@ This is called optimization, but how does optimization actually work?
 Gradient descent is a widely used optimization algorithm, most other optimization algorithms are based on it.
 It works as follows: Imagine a neural network with only one neuron.
 Take a look at the figure below. The plot shows the loss as a function of the weight of the neuron.
+
+![](fig/03_gradient_descent.png){alt='Plot of the loss as a function of the weights. Through gradient descent the global loss minimum is found'}
+
 As you can see there is a global loss minimum, we would like to find the weight at this point in the parabola.
 To do this, we initialize the model weight with some random value. Then we compute the gradient of the loss function with respect
 to the weight. This tells us how much the loss function will change if we change the weight by a small amount.
 Then, we update the weight by taking a small step in the direction of the negative gradient, so down the slope.
 This will slightly decrease the loss. This process is repeated until the loss function reaches a minimum.
 The size of the step that is taken in each iteration is called the 'learning rate'.
-
-![](fig/03_gradient_descent.png){alt='Plot of the loss as a function of the weights. Through gradient descent the global loss minimum is found'}
 
 ### Batch gradient descent
 You could use the entire training dataset to perform one learning step in gradient descent,
@@ -407,7 +408,7 @@ def plot_history(history, metrics):
 plot_history(history, 'root_mean_squared_error')
 ```
 
-![](fig/03_training_history_1_rmse.png){alt='Plot of the RMSE over epochs for the trained model that shows a decreasing error metric'}
+![](fig/03_training_history_1_rmse.png){alt='Plot of the RMSE over epochs for the trained model that shows a decreasing error metric.'}
 
 This looks very promising! Our metric ("RMSE") is dropping nicely and while it maybe keeps fluctuating a bit it does end up at fairly low *RMSE* values.
 But the *RMSE* is just the root *mean* squared error, so we might want to look a bit more in detail how well our just trained model does in predicting the sunshine hours.
@@ -445,7 +446,7 @@ def plot_predictions(y_pred, y_true, title):
 plot_predictions(y_train_predicted, y_train, title='Predictions on the training set')
 ```
 
[…]
