### 2025-09-01 8883bb0
markdown source builds
Auto-generated via `{sandpaper}`
Source  : e518b333c4604b85a2abe4d001f6ef76fb14ae2c
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-01 12:42:10 +0000
Message : Merge pull request #594 from carpentries-lab/change-confusion-matrix-cmap

Change colormap of confustion matrix to Blues
