### 2025-09-01 7146312
site deploy
Auto-generated via `{sandpaper}`
Source  : 8883bb04e4890975cc1b7afc4ab3ffabc9b67135
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-01 12:43:03 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : e518b333c4604b85a2abe4d001f6ef76fb14ae2c
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-01 12:42:10 +0000
Message : Merge pull request #594 from carpentries-lab/change-confusion-matrix-cmap

Change colormap of confustion matrix to Blues
