### 2025-09-01 d74ced4
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 557699c96be7c8dd998c1f47d21f51957e51d07b
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-01 12:43:17 +0000
Message : Merge pull request #596 from carpentries-lab/fix/typos_plots_547

Fix/typos plots 547
