+++ md5sum.txt
@@ -7,7 +7,7 @@
 "workshops.md" "ba594ced8acc714eb6785d1c63191bb8" "site/built/workshops.md" "2025-05-07"
 "episodes/1-introduction.md" "8dabfa4853b660c8bfcb0aea5f435029" "site/built/1-introduction.md" "2025-02-11"
 "episodes/2-keras.md" "6c8b4f60b6ea24f6eee050df635d3789" "site/built/2-keras.md" "2025-09-01"
-"episodes/3-monitor-the-model.md" "4be34c9bbba665390860faa71ccc22cc" "site/built/3-monitor-the-model.md" "2025-09-01"
+"episodes/3-monitor-the-model.md" "7240f0fc522e2c5817efafe02898ce37" "site/built/3-monitor-the-model.md" "2025-09-01"
 "episodes/4-advanced-layer-types.md" "61bb5fc50d1b2cde8163fdfff1a99bb0" "site/built/4-advanced-layer-types.md" "2025-09-01"
 "episodes/5-transfer-learning.md" "9dae01295f9f0a566a71c05672cc052f" "site/built/5-transfer-learning.md" "2025-09-01"
 "episodes/6-outlook.md" "007728216562f3b52b983ff1908af5b7" "site/built/6-outlook.md" "2025-02-11"
