### 2025-09-01 5ee62f4
site deploy
Auto-generated via `{sandpaper}`
Source  : d74ced4c473f7986385dbd8cb03568bc4a739507
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-01 12:44:11 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : 557699c96be7c8dd998c1f47d21f51957e51d07b
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-01 12:43:17 +0000
Message : Merge pull request #596 from carpentries-lab/fix/typos_plots_547

Fix/typos plots 547
