### 2025-09-01 234f5ac
expand alt text descriptions
