### 2025-09-01 e78fe3c
fix challenge-solution format
