+++ episodes/4-advanced-layer-types.md
@@ -100,7 +100,7 @@ The remainder of the shape, namely `(64, 64, 3)`, denotes
 the dimension of one image. The last value 3 is typical for color images,
 and stands for the three color channels **R**ed, **G**reen, **B**lue.
 
-::: challenge
+:::: challenge
 
 ### Number of features in Dollar Street 10
 
@@ -112,12 +112,12 @@ How many features does one image in the Dollar Street 10 dataset have?
 - D. 878
 
 
-:::: solution
+::: solution
 The correct solution is C: 12288
 
 There are 4096 pixels in one image (64 * 64), each pixel has 3 channels (RGB). So 4096 * 3 = 12288.
-::::
 :::
+::::
 
 
 We can find out the range of values of our input data as follows:
@@ -165,7 +165,7 @@ In the previous episodes, we used 'fully connected layers' , that connected all 
 This results in many connections, and thus many weights to be learned, in the network.
 Note that our input dimension is now quite high (even with small pictures of `64x64` pixels): we have 12288 features.
 
-::: challenge
+:::: challenge
 ## Number of parameters{#parameters-exercise-1}
 Suppose we create a single Dense (fully connected) layer with 100 hidden units that connect to the input pixels, how many parameters does this layer have?
 
@@ -174,7 +174,7 @@ Suppose we create a single Dense (fully connected) layer with 100 hidden units t
 - C. 100
 - D. 12288
 
-:::: solution
+::: solution
 ## Solution
 The correct answer is B: Each entry of the input dimensions, i.e. the `shape` of one single data point, is connected with 100 neurons of our hidden layer, and each of these neurons has a bias term associated to it. So we have `1228900` parameters to learn.
 ```python
@@ -212,8 +212,8 @@ Model: "functional"
 
  Non-trainable params: 0 (0.00 B)
 ```
-::::
 :::
+::::
 
 We can decrease the number of units in our hidden layer, but this also decreases the number of patterns our network can remember. Moreover, if we increase the image size, the number of weights will 'explode', even though the task of recognizing large images is not necessarily more difficult than the task of recognizing small images.
 
@@ -248,11 +248,11 @@ there that enable users to interactively play around with images and convolution
      shows animated examples of the different components of convolutional neural nets 
 :::
 
-::: challenge
+:::: challenge
 ## Border pixels
 What, do you think, happens to the border pixels when applying a convolution?
 
-:::: solution
+::: solution
 ## Solution
 There are different ways of dealing with border pixels.
 You can ignore them, which means that your output image is slightly smaller then your input.
@@ -262,18 +262,18 @@ In that case, the output image will have the same size as the input image.
 [This callout in the Data Carpentry: Image Processing with Python curriculum](https://datacarpentry.org/image-processing/06-blurring.html#callout4)
 provides more detail about convolution at the boundaries of an image,
 in the context of applying a _Gaussian blur_.
-::::
 :::
+::::
 
-::: challenge
+:::: challenge
 ## Number of model parameters
 Suppose we apply a convolutional layer with 100 kernels of size 3 * 3 * 3 (the last dimension applies to the rgb channels) to our images of 32 * 32 * 3 pixels. How many parameters do we have? Assume, for simplicity, that the kernels do not use bias terms. Compare this to the answer of the earlier exercise, ["Number of Parameters"](#parameters-exercise-1).
 
-:::: solution
+::: solution
 ## Solution
[…]
