### 2025-09-01 39864cb
Add alt text for updated plot.
