### 2025-09-01 8a88503
fix typos and address inaccuracy
