+++ md5sum.txt
@@ -5,12 +5,12 @@
 "index.md" "8b5609014b8028029f48266bc751663e" "site/built/index.md" "2025-05-06"
 "links.md" "8184cf4149eafbf03ce8da8ff0778c14" "site/built/links.md" "2025-02-11"
 "workshops.md" "ba594ced8acc714eb6785d1c63191bb8" "site/built/workshops.md" "2025-05-07"
-"episodes/1-introduction.md" "8dabfa4853b660c8bfcb0aea5f435029" "site/built/1-introduction.md" "2025-02-11"
-"episodes/2-keras.md" "6c8b4f60b6ea24f6eee050df635d3789" "site/built/2-keras.md" "2025-09-01"
-"episodes/3-monitor-the-model.md" "7240f0fc522e2c5817efafe02898ce37" "site/built/3-monitor-the-model.md" "2025-09-01"
-"episodes/4-advanced-layer-types.md" "61bb5fc50d1b2cde8163fdfff1a99bb0" "site/built/4-advanced-layer-types.md" "2025-09-01"
-"episodes/5-transfer-learning.md" "9dae01295f9f0a566a71c05672cc052f" "site/built/5-transfer-learning.md" "2025-09-01"
-"episodes/6-outlook.md" "007728216562f3b52b983ff1908af5b7" "site/built/6-outlook.md" "2025-02-11"
+"episodes/1-introduction.md" "42be5910844ee2c2c8ab7c3fbf9791f1" "site/built/1-introduction.md" "2025-09-01"
+"episodes/2-keras.md" "c0709c14478f22a34169df630aca5957" "site/built/2-keras.md" "2025-09-01"
+"episodes/3-monitor-the-model.md" "dc6a1e2b29983b2ec1862b22a79324de" "site/built/3-monitor-the-model.md" "2025-09-01"
+"episodes/4-advanced-layer-types.md" "92141a294e5272c706dd0269e25f4bd7" "site/built/4-advanced-layer-types.md" "2025-09-01"
+"episodes/5-transfer-learning.md" "b85308f5b2d0a65cb86209f21252e496" "site/built/5-transfer-learning.md" "2025-09-01"
+"episodes/6-outlook.md" "84fb67b1725bf91e42ba668fb75c1c12" "site/built/6-outlook.md" "2025-09-01"
 "instructors/bonus-material.md" "382832ea4eb097fc7781cb36992c1955" "site/built/bonus-material.md" "2025-02-11"
 "instructors/design.md" "1537f9d0c90cdfdb1781bd3daf94dadd" "site/built/design.md" "2025-02-11"
 "instructors/instructor-notes.md" "541a3ccba7b1edff5040ebfb5a1b0c3d" "site/built/instructor-notes.md" "2025-07-18"
