### 2025-09-01 49c3861
site deploy
Auto-generated via `{sandpaper}`
Source  : 099a1b10220f18f5a726a1ba763deab72df88e85
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-01 13:54:02 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : b5d6ae85227f44a6b913e8b0e593e71da045ad3c
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-01 13:53:08 +0000
Message : Merge pull request #600 from carpentries-lab/569-indentation-of-nested-challengesolution-blocks-is-incorrect

fix challenge-solution format
