### 2025-09-02 4b981af
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 54dd6ea8b13ccb42d84af3a88af3c1b9a02828b8
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 07:52:15 +0000
Message : Merge pull request #582 from UW-Madison-DataScience/episode3_add-diag-ref-line

Add diagonal ref. line to plot_predictions() function and update figs
