+++ md5sum.txt
@@ -7,7 +7,7 @@
 "workshops.md" "ba594ced8acc714eb6785d1c63191bb8" "site/built/workshops.md" "2025-05-07"
 "episodes/1-introduction.md" "42be5910844ee2c2c8ab7c3fbf9791f1" "site/built/1-introduction.md" "2025-09-01"
 "episodes/2-keras.md" "c0709c14478f22a34169df630aca5957" "site/built/2-keras.md" "2025-09-01"
-"episodes/3-monitor-the-model.md" "dc6a1e2b29983b2ec1862b22a79324de" "site/built/3-monitor-the-model.md" "2025-09-01"
+"episodes/3-monitor-the-model.md" "ab8fb8e73c22e962c03562e1fe834f0e" "site/built/3-monitor-the-model.md" "2025-09-02"
 "episodes/4-advanced-layer-types.md" "92141a294e5272c706dd0269e25f4bd7" "site/built/4-advanced-layer-types.md" "2025-09-01"
 "episodes/5-transfer-learning.md" "b85308f5b2d0a65cb86209f21252e496" "site/built/5-transfer-learning.md" "2025-09-01"
 "episodes/6-outlook.md" "84fb67b1725bf91e42ba668fb75c1c12" "site/built/6-outlook.md" "2025-09-01"
