### 2025-09-02 84eefd0
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 78c4089c7af4c581b0cbd0ef25e9db12eac40da0
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 07:52:57 +0000
Message : Merge pull request #595 from carpentries-lab/550-make-create_nn-consistent-with-episode-3

create_nn() now takes the input shape as an argument
