### 2025-09-02 d8ed3b9
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 8d8473715ecd2ac9bdaee39212762aa84e6d14a6
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 07:53:56 +0000
Message : Merge pull request #601 from carpentries-lab/bonus-alt-text

improve alt text description
