+++ md5sum.txt
@@ -11,7 +11,7 @@
 "episodes/4-advanced-layer-types.md" "f16e2d9494970bc706bcd854279f8cd7" "site/built/4-advanced-layer-types.md" "2025-09-02"
 "episodes/5-transfer-learning.md" "b85308f5b2d0a65cb86209f21252e496" "site/built/5-transfer-learning.md" "2025-09-01"
 "episodes/6-outlook.md" "84fb67b1725bf91e42ba668fb75c1c12" "site/built/6-outlook.md" "2025-09-01"
-"instructors/bonus-material.md" "382832ea4eb097fc7781cb36992c1955" "site/built/bonus-material.md" "2025-02-11"
+"instructors/bonus-material.md" "db1b33b9c227e2b312d084c532c9fe74" "site/built/bonus-material.md" "2025-09-02"
 "instructors/design.md" "1537f9d0c90cdfdb1781bd3daf94dadd" "site/built/design.md" "2025-02-11"
 "instructors/instructor-notes.md" "541a3ccba7b1edff5040ebfb5a1b0c3d" "site/built/instructor-notes.md" "2025-07-18"
 "instructors/schedule.md" "332b32d24f144b29a280176e6b5d015f" "site/built/schedule.md" "2025-03-10"
