### 2025-09-02 a31ba8b
markdown source builds
Auto-generated via `{sandpaper}`
Source  : bf53ed3e4ef51bfabf4aece2cbc0902102883703
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 07:54:41 +0000
Message : Merge pull request #585 from UW-Madison-DataScience/Episode5_fix-bulleted-list

fix bulleted list formatting
