+++ md5sum.txt
@@ -9,7 +9,7 @@
 "episodes/2-keras.md" "c0709c14478f22a34169df630aca5957" "site/built/2-keras.md" "2025-09-01"
 "episodes/3-monitor-the-model.md" "ab8fb8e73c22e962c03562e1fe834f0e" "site/built/3-monitor-the-model.md" "2025-09-02"
 "episodes/4-advanced-layer-types.md" "f16e2d9494970bc706bcd854279f8cd7" "site/built/4-advanced-layer-types.md" "2025-09-02"
-"episodes/5-transfer-learning.md" "b85308f5b2d0a65cb86209f21252e496" "site/built/5-transfer-learning.md" "2025-09-01"
+"episodes/5-transfer-learning.md" "553443c26e6db723ac1e5eb34e6533d5" "site/built/5-transfer-learning.md" "2025-09-02"
 "episodes/6-outlook.md" "84fb67b1725bf91e42ba668fb75c1c12" "site/built/6-outlook.md" "2025-09-01"
 "instructors/bonus-material.md" "db1b33b9c227e2b312d084c532c9fe74" "site/built/bonus-material.md" "2025-09-02"
 "instructors/design.md" "1537f9d0c90cdfdb1781bd3daf94dadd" "site/built/design.md" "2025-02-11"
