### 2025-09-02 2f47d67
site deploy
Auto-generated via `{sandpaper}`
Source  : a31ba8b6bad5d90036b6bb3a3f0dd752a5a866f7
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-02 07:55:30 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : bf53ed3e4ef51bfabf4aece2cbc0902102883703
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 07:54:41 +0000
Message : Merge pull request #585 from UW-Madison-DataScience/Episode5_fix-bulleted-list

fix bulleted list formatting
