### 2025-09-02 d85da8c
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 07d5a6f8d9f0e81fb1ba02665c1e8f390476988b
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 07:55:07 +0000
Message : Merge pull request #586 from UW-Madison-DataScience/episode3-updates

add missing import to pandas (if downloading data from zenodo)
