### 2025-09-02 34a1c4b
markdown source builds
Auto-generated via `{sandpaper}`
Source  : abf98fff5becdef88c1c08d572c1d196a9b4c926
Branch  : main
Author  : Toby Hodges <tobyhodges@carpentries.org>
Time    : 2025-09-02 08:06:47 +0000
Message : Merge pull request #598 from carpentries-lab/ep4-alt-text

Expand alt text descriptions in Episode 4
