### 2025-09-02 e634a3a
site deploy
Auto-generated via `{sandpaper}`
Source  : 34a1c4b9d770ed6bffbfba94dfcc9c06b15f0bb8
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-02 08:07:37 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : abf98fff5becdef88c1c08d572c1d196a9b4c926
Branch  : main
Author  : Toby Hodges <tobyhodges@carpentries.org>
Time    : 2025-09-02 08:06:47 +0000
Message : Merge pull request #598 from carpentries-lab/ep4-alt-text

Expand alt text descriptions in Episode 4
