### 2025-09-02 8b7cd70
correct image dimensions
