### 2025-09-02 f37f936
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 4dca474f5d4832b750c873eb0247f18f5c300de5
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 08:29:00 +0000
Message : Merge pull request #603 from carpentries-lab/549-image-size

Correct image dimensions
