### 2025-09-02 780683e
site deploy
Auto-generated via `{sandpaper}`
Source  : f37f93618253d2e2c64d75d2472380b85445e639
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-02 08:29:46 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : 4dca474f5d4832b750c873eb0247f18f5c300de5
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 08:29:00 +0000
Message : Merge pull request #603 from carpentries-lab/549-image-size

Correct image dimensions
