### 2025-09-02 1935a9f
site deploy
Auto-generated via `{sandpaper}`
Source  : a578d81de81bd2f1a096036015764f1a7fa91a37
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-02 08:34:20 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : ed61d60502609b3ac9b3f714fcb990a08a7abcf7
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 08:33:10 +0000
Message : Merge pull request #604 from carpentries-lab/563-d2l

add Dive into Deep Learning to list of resources
