+++ md5sum.txt
@@ -16,7 +16,7 @@
 "instructors/instructor-notes.md" "541a3ccba7b1edff5040ebfb5a1b0c3d" "site/built/instructor-notes.md" "2025-07-18"
 "instructors/schedule.md" "332b32d24f144b29a280176e6b5d015f" "site/built/schedule.md" "2025-03-10"
 "instructors/survey-templates.md" "ea5d46e7b54d335f79e57a7bc31d1c5c" "site/built/survey-templates.md" "2025-02-11"
-"learners/reference.md" "e47218673643f23431c540c2d3b27868" "site/built/reference.md" "2025-02-11"
+"learners/reference.md" "8f72ebcd8c085f01e39072e685272de3" "site/built/reference.md" "2025-09-02"
 "learners/setup.md" "2e741a6d76091da5c832ab010380f0e1" "site/built/setup.md" "2025-09-01"
 "paper/paper.md" "977a3c4badad0bae31972e6e86045bde" "site/built/paper.md" "2025-09-01"
 "profiles/learner-profiles.md" "ef0f26dd0874387d80ed3fd468b99e23" "site/built/learner-profiles.md" "2025-02-11"
