### 2025-09-02 ad93808
Add Callout about metrics.
