### 2025-09-02 b0b066e
markdown source builds
Auto-generated via `{sandpaper}`
Source  : ba158ebe602ebc63ee83d9747901319ee4b17ce6
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 09:33:37 +0000
Message : Merge pull request #606 from carpentries-lab/578-why-is-it-always-none-in-the-layer-shape

added clarification comment about None value in params
