### 2025-09-02 7c33e20
site deploy
Auto-generated via `{sandpaper}`
Source  : b0b066ec80880d4338d931fd0d5a0d11c55e1b42
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-02 09:36:53 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : ba158ebe602ebc63ee83d9747901319ee4b17ce6
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 09:33:37 +0000
Message : Merge pull request #606 from carpentries-lab/578-why-is-it-always-none-in-the-layer-shape

added clarification comment about None value in params
