### 2025-09-02 9dd15ef
markdown source builds
Auto-generated via `{sandpaper}`
Source  : ad938618383de33ba6e612aaf658b0ec0d0468f6
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 09:34:31 +0000
Message : Merge pull request #599 from carpentries-lab/fix/545_plot

Update training curve plot.
