+++ md5sum.txt
@@ -6,7 +6,7 @@
 "links.md" "8184cf4149eafbf03ce8da8ff0778c14" "site/built/links.md" "2025-02-11"
 "workshops.md" "ba594ced8acc714eb6785d1c63191bb8" "site/built/workshops.md" "2025-05-07"
 "episodes/1-introduction.md" "42be5910844ee2c2c8ab7c3fbf9791f1" "site/built/1-introduction.md" "2025-09-01"
-"episodes/2-keras.md" "c7596505ad92bbfa26c6e8d664088a7a" "site/built/2-keras.md" "2025-09-02"
+"episodes/2-keras.md" "0b10aa86b79a6d0e2156cb01e3751b6d" "site/built/2-keras.md" "2025-09-02"
 "episodes/3-monitor-the-model.md" "8eda70a03e5225033ba05563b517ec88" "site/built/3-monitor-the-model.md" "2025-09-02"
 "episodes/4-advanced-layer-types.md" "bc997ca7370fdd7a49ebff3a967738bd" "site/built/4-advanced-layer-types.md" "2025-09-02"
 "episodes/5-transfer-learning.md" "553443c26e6db723ac1e5eb34e6533d5" "site/built/5-transfer-learning.md" "2025-09-02"
