### 2025-09-02 f7727b3
Update 'Choosing a Metric' callout.
