### 2025-09-02 cdffe04
Add config and Github action for Linkspector.
