+++ .github/workflows/linkspector.yml
@@ -1,0 +1,10 @@
+name: Linkspector
+on: [pull_request]
+jobs:
+  check-links:
+    name: runner / linkspector
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v5
+      - name: Run linkspector
+        uses: umbrelladocs/action-linkspector@v1
