### 2025-09-02 fe46121
site deploy
Auto-generated via `{sandpaper}`
Source  : 0998ef60b60349784988748e90f5244a53213d97
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-02 11:09:56 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : 71cb4087d37fd22a8d4d00bcbd395354f9b1a7d7
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-02 11:06:37 +0000
Message : Merge pull request #609 from carpentries-lab/explain-need-for-test-set

Clarify the solution which describes the need for a test set
