+++ md5sum.txt
@@ -8,7 +8,7 @@
 "episodes/1-introduction.md" "42be5910844ee2c2c8ab7c3fbf9791f1" "site/built/1-introduction.md" "2025-09-01"
 "episodes/2-keras.md" "ed541622e363403249f62c0b10794fdb" "site/built/2-keras.md" "2025-09-02"
 "episodes/3-monitor-the-model.md" "8eda70a03e5225033ba05563b517ec88" "site/built/3-monitor-the-model.md" "2025-09-02"
-"episodes/4-advanced-layer-types.md" "bc997ca7370fdd7a49ebff3a967738bd" "site/built/4-advanced-layer-types.md" "2025-09-02"
+"episodes/4-advanced-layer-types.md" "ca13fa5d8097ff098019ae8e106efe87" "site/built/4-advanced-layer-types.md" "2025-09-02"
 "episodes/5-transfer-learning.md" "553443c26e6db723ac1e5eb34e6533d5" "site/built/5-transfer-learning.md" "2025-09-02"
 "episodes/6-outlook.md" "84fb67b1725bf91e42ba668fb75c1c12" "site/built/6-outlook.md" "2025-09-01"
 "instructors/bonus-material.md" "db1b33b9c227e2b312d084c532c9fe74" "site/built/bonus-material.md" "2025-09-02"
