### 2025-09-02 #281 @magsol
Hi @carschno! I'll see how quickly we can get things moving here 🙂

Thanks to @labarba laying out the guidelines for how [expedited reviews from Carpentries](https://github.com/openjournals/jose-reviews/issues/281#issuecomment-3114563867) work, it looks like we need only a single reviewer to do a quick pass. I always like to ask the author(s) for their suggestions, and I'm happy to run with those. All I need to know is how to reach out to them! Let me know and I can send invites out immediately.
