### 2025-09-03 74088c9
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 50ec9f2e44169cba4f5d7f3dfbd29004b9fefebb
Branch  : main
Author  : José Angel Daza <josdaza.a@gmail.com>
Time    : 2025-09-03 07:36:00 +0000
Message : Merge pull request #610 from carpentries-lab/explain-nb-params-cnn

Demystify the number of parameters in Conv2D layers
