+++ md5sum.txt
@@ -6,7 +6,7 @@
 "links.md" "8184cf4149eafbf03ce8da8ff0778c14" "site/built/links.md" "2025-02-11"
 "workshops.md" "ba594ced8acc714eb6785d1c63191bb8" "site/built/workshops.md" "2025-05-07"
 "episodes/1-introduction.md" "42be5910844ee2c2c8ab7c3fbf9791f1" "site/built/1-introduction.md" "2025-09-01"
-"episodes/2-keras.md" "ed541622e363403249f62c0b10794fdb" "site/built/2-keras.md" "2025-09-02"
+"episodes/2-keras.md" "ab5ebd62fd3e6cc2ad69bd805d26f7f1" "site/built/2-keras.md" "2025-09-03"
 "episodes/3-monitor-the-model.md" "8eda70a03e5225033ba05563b517ec88" "site/built/3-monitor-the-model.md" "2025-09-02"
 "episodes/4-advanced-layer-types.md" "eaca1f96ead3140467a4ba7f069284fa" "site/built/4-advanced-layer-types.md" "2025-09-03"
 "episodes/5-transfer-learning.md" "553443c26e6db723ac1e5eb34e6533d5" "site/built/5-transfer-learning.md" "2025-09-02"
@@ -16,7 +16,7 @@
 "instructors/instructor-notes.md" "541a3ccba7b1edff5040ebfb5a1b0c3d" "site/built/instructor-notes.md" "2025-07-18"
 "instructors/schedule.md" "332b32d24f144b29a280176e6b5d015f" "site/built/schedule.md" "2025-03-10"
 "instructors/survey-templates.md" "ea5d46e7b54d335f79e57a7bc31d1c5c" "site/built/survey-templates.md" "2025-02-11"
-"learners/reference.md" "8f72ebcd8c085f01e39072e685272de3" "site/built/reference.md" "2025-09-02"
+"learners/reference.md" "2ff8ad45ffba9a6603b89340c4879d94" "site/built/reference.md" "2025-09-03"
 "learners/setup.md" "2e741a6d76091da5c832ab010380f0e1" "site/built/setup.md" "2025-09-01"
 "paper/paper.md" "977a3c4badad0bae31972e6e86045bde" "site/built/paper.md" "2025-09-01"
 "profiles/learner-profiles.md" "ef0f26dd0874387d80ed3fd468b99e23" "site/built/learner-profiles.md" "2025-02-11"
