### 2025-09-03 #281 @magsol
Ahh, I did not realize they were already listed in the JOSE reviewers spreadsheet. Thank you @carschno! I'll reach out to them, and we'll go from there. Hopefully, it shouldn't take me more than a week (response time notwithstanding).
