### 2025-09-04 #281 @magsol
Invite status (will update as I continue to hear back):

- Brandon LeBeau: declined (9/4)
- Manjari Narayan: email bounced
- Abhinav Maurya:
- Gajendra Deshpande:
- Albert Kim: declined (9/6)
- Kelly Bodwin:
