### 2025-09-09 a20596b
site deploy
Auto-generated via `{sandpaper}`
Source  : cb3ba1b25abb5d802629a1de607664103507c10f
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-03 11:49:14 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : 1d44463e4454980bb2dd9a000c9d49ecf5011f49
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-03 11:48:33 +0000
Message : Merge pull request #608 from carpentries-lab/fix/links

Add Linkspector
