### 2025-09-11 #307 @editorialbot
[REVIEW]: Introduction to deep learning: Carpentries-style hands-on lesson material for introducing researchers to deep learning
**Submitting author:** @svenvanderburg (<a href="http://orcid.org/0000-0003-1250-6968">Sven A. van der Burg</a>)
**Repository:** https://github.com/carpentries-lab/deep-learning-intro
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.2
