### 2025-09-11 #307 @manideepikaa
Thanks @magsol  for giving me the opportunity to review the  paper,
The topic was well articulated and documented .
However, these were few of my recommendations while reviewing the paper.

1. A more stronger and straight-forward statement of need would be great.
2. For the requirements for the software submission, It's Ok , as package.json file was missing for a good Submission.
3. The Github version number was missing and the release version number in the code.
4. In terms of Authorship which involves the visible contributions made by the author to the module, Could you please elaborate on the contributions of Anne Fouilloux as i could see only 2 commits from the author towards the code.
5. The paper is missing a story on how the authors developed it and their expertise.
