### 2025-09-11 #281 @carschno
>  They're a former graduate student of mine who is expert in deep learning. I've every confidence they'll do a great job.

That sounds fantastic, I would be happy to move on with them!
