### 2025-09-12 #307 @carschno
@manideepikaa thanks a lot for the quick review!
FYI: as Sven has left the eScience Center a while ago, I will try to act as a contact person for the review (see https://github.com/openjournals/jose-reviews/issues/281#issuecomment-2853374567), and to address the issues you have found.
