### 2025-09-12 #307 @carschno
> In terms of Authorship which involves the visible contributions made by the author to the module, Could you please elaborate on the contributions of Anne Fouilloux as i could see only 2 commits from the author towards the code.

That is a good point! I genuinely do not know some of the listed authors and what they contributed before I joined, so I find it hard to refine the author list. @ashwinvis should be added anyway, he has joined last year.

Regarding @annefou, I am a bit lost myself. She is listed with one commit of 347 lines in the [contributors list](https://github.com/carpentries-lab/deep-learning-intro/graphs/contributors), but there is no commit of hers in the [commit history](https://github.com/carpentries-lab/deep-learning-intro/commits?author=annefou). Weirdly, exactly the same shows for @dafnevk, including the exact number of lines.

There are a couple of more people who seem to be missing, and I don't want to remove authors due to lack of my own knowledge. Before I go through the entire list, perhaps @tobyhodges do you have a good overview of active authors, and/or some practical guidelines for determining authorship. Also, what are the standards regarding co-author ordering?
