### 2025-09-12 #307 @tobyhodges
Hi, yes. @annefou was heavily involved in the early stages of discussion and design of the lesson. She is not represented in much of the commit history but is partially responsible for the conceptualisation of the lesson. (See [this comment for some evidence of that](https://github.com/carpentries-lab/deep-learning-intro/pull/60#pullrequestreview-589477092), please let me know if you need more than that.)

@dafnevk coordinated the project for a while from the Netherlands eScience Center, leading coworking calls, contributing to discussions, and writing a lot of the content. I am not sure why that isn't represented in the contributions page, but a grep on the `git log` of `main` shows 136 commits:

```zsh
(base) deep-learning-intro (main) » git log | grep -c 'Dafne van Kuppevelt'
136
```

If you take a look through the history of closed issues and PRs you will quickly discover plenty of evidence of Dafne's crucial contributions.

@manideepikaa thanks very much for checking. I really appreciate your diligence 🙌 Please let me know if I can provide any more information.
