### 2025-09-13 #307 @carschno
> Would you guys consider reordering the co-authors or you were okay with the present ordering ?

We would definitely consider that, along with adding other authors, at least @ashwinvis as suggested [above](https://github.com/openjournals/jose-reviews/issues/307#issuecomment-3284166348).

Would you have a good guideline for which authors to list at all, and in which order?
