### 2025-09-16 #307 @ashwinvis
@carschno the guidelines for authorship are listed here: https://openjournals.readthedocs.io/en/jose/submitting.html#authorship
