### 2025-09-16 #307 @ashwinvis
Regarding authorship, thank you for considering including me. While the discussion is going on, I noticed that my colleague @wikfeldt was also listed with significant number of commits.

    16      Kjartan Thor Wikfeldt

In addition to the commits, in 2022, he had also actively contributed in providing feedback to the lessons. See here:

https://github.com/carpentries-lab/deep-learning-intro/issues?q=is%3Aissue%20state%3Aclosed%20author%3Awikfeldt

Since the review has started, please consider him also to include amongst the authors.
