### 2025-09-16 #307 @carschno
> [@carschno](https://github.com/carschno) the guidelines for authorship are listed here: https://openjournals.readthedocs.io/en/jose/submitting.html#authorship

I was not aware of that, thanks!
However, it leaves two questions open:

1. Carpentries lessons have a special process (see [this comment](https://github.com/openjournals/jose-reviews/issues/281#issuecomment-3114563867)), so perhaps there are also special author policies regarding lesson authorship vs. JOSE article authorship
2. Which author ordering is recommended?
