### 2025-09-16 #307 @magsol
@carschno

Regarding 1, there are no special authorship considerations that I'm aware of in the context of Carpentries lessons being fast-tracked at JOSE (unless @labarba knows of any?). So I'd say, whatever order you determined and which was approved by Carpentries, should be sufficient for JOSE as well (as long as the single explicit guideline listed in JOSE's authorship document holds: that someone who only contributed financially or is simply in the same organization cannot be named an author).

Aside from that single stipulation, whatever you had for Carpentries should be sufficient here, though I see no reason why you couldn't also add people if you wanted.
