### 2025-09-23 #307 @carschno
@ashwinvis and @wikfeldt could you add yourselves to the author list in [the paper](https://github.com/carpentries-lab/deep-learning-intro/blob/main/paper/paper.md?plain=1), and create a PR? I think that would be the best way to ensure your affiliations etc. are correct, and to formally ensure your consent to being in the author list.
