### 2025-09-23 #307 @wikfeldt
thanks for the ping, i've now added myself and Ashwin in https://github.com/carpentries-lab/deep-learning-intro/pull/611
