### 2025-09-23 #307 @carschno
> Would you guys consider reordering the co-authors or you were okay with the present ordering ?

Done, the authors list should now be finalized.
