### 2025-09-30 93f1ab5
site deploy
Auto-generated via `{sandpaper}`
Source  : 8b0c71f9fda61ed81689f4c207da16fe2d8f1ceb
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-09-23 17:59:13 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : fe4a28744fa77e66a89b00357e77a765bb7ae887
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-09-23 17:58:28 +0000
Message : Merge pull request #611 from wikfeldt/patch-1

add Thor and Ashwin to author list
