### 2025-10-05 #307 @carschno
EDIT: I had missed your updated checklist [above](https://github.com/openjournals/jose-reviews/issues/307#issuecomment-3282171720), I will work on that.
