### 2025-10-06 266e3a5
markdown source builds
Auto-generated via `{sandpaper}`
Source  : d4dab41b3707772075979894b70b6f3fc098c704
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-10-06 08:48:43 +0000
Message : Add Lesson Development section.
