### 2025-10-06 faa6ed9
site deploy
Auto-generated via `{sandpaper}`
Source  : 266e3a5a28129246803cdcd8df8340b5f1e633b0
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-10-06 08:49:49 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : d4dab41b3707772075979894b70b6f3fc098c704
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-10-06 08:48:43 +0000
Message : Add Lesson Development section.
