### 2025-10-06 #307 @carschno
@manideepikaa I have updated the paper, hopefully addressing the points you indicated in your checklist.

Regarding the versioning, I have created a new release (`v1.2.2`) of the lesson. Apparently, you as an editor have to trigger the update with this command:

> @editorialbot set <v1.2.2> as version
