### 2025-10-14 #307 @manideepikaa
@carschno
Could you please address these as well.

        1. For the requirements for the software submission, It's Ok , as package.json file was missing for a good Submission.
        2. The Github version number was missing and the release version number in the code.
        3. The paper is missing a story on how the authors developed it and their expertise.
