### 2025-10-15 #307 @carschno
Hi @manideepikaa,
I think I missed these points, because I only checked your [checklist](https://github.com/openjournals/jose-reviews/issues/307#issuecomment-3282171720).

Per you specific points:

>
>     1. For the requirements for the software submission, It's Ok , as package.json file was missing for a good Submission.
>     2. The Github version number was missing and the release version number in the code.

I suspect that these specific points in the guidelines are not really applicable for a lesson submission. A `package.json` file for dependency management does not quite make sense here, as we do not provide software. It does not seem to exist in the [other lessons](https://github.com/orgs/carpentries-lab/repositories?type=all) either.
The same applies to the Github version number in the code -- we do not provide code.

>     3. The paper is missing a story on how the authors developed it and their expertise.>

I have added a new section to address this issue while trying not to interfere with the paper too much because I am not the first author. This is the relevant commit: https://github.com/carpentries-lab/deep-learning-intro/commit/d4dab41b3707772075979894b70b6f3fc098c704

If this does not satisfy your requirements about the lesson development story, could you provide some feedback on how to improve?
