### 2025-10-16 #307 @manideepikaa
@carschno , I agree with your points
for #1 and #2, as i did noticed a couple of code files, i did brought this up.
As the main goal here was lesson submission than code,  I am fine with that.
for #3, the story development, My idea was to emphasize more on the expertise of the authors
I completely understand that there are multiple authors involved, So, it makes sense .
