### 2025-10-16 #307 @magsol
Ok, this looks great! Let go through a few steps here...
