### 2025-10-16 #307 @magsol
@carschno Hi, I just want to double-check that v1.2.2 (the last release in your repo, and the archived version on Zenodo) is indeed the latest version that we are going to publish. I'm just double-checking since this version came out Oct 6 and didn't want to miss any changes that may have been made since then :)
