### 2025-10-17 #307 @carschno
> for [#3](https://github.com/openjournals/jose-reviews/issues/3), the story development, My idea was to emphasize more on the expertise of the authors
> I completely understand that there are multiple authors involved, So, it makes sense .

Thanks for pointing this out! You are right that there is little emphasis on this aspect, but I also agree that this might not be feasible/necessary in this case. I will leave it as-is then.
