### 2025-10-17 #307 @carschno
> [@carschno](https://github.com/carschno) Hi, I just want to double-check that v1.2.2 (the last release in your repo, and the archived version on Zenodo) is indeed the latest version that we are going to publish. I'm just double-checking since this version came out Oct 6 and didn't want to miss any changes that may have been made since then :)

I can confirm that there have no changes been made to the lesson since then.

Just to elaborate: the paper itself resides in the same repository, and has been updated after the last release. Anyway, changes to the paper only do not justify a new lesson release, so the lesson version remains 1.2.2.
