### 2025-10-19 #307 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
