### 2025-10-20 #307 @carschno
Hi @magsol and @manideepikaa,
thank you both so much for pushing this forward!

During a final proofread, I have noticed a tiny remaining issue: the survey data in Table 2 refers to the CIFAR-10 dataset only, but should also refer to the Dollarstreet-10 dataset, which has been used in the more recent iterations of the workshop.

I have pushed a slightly updated version of `paper.md` (https://github.com/carpentries-lab/deep-learning-intro/commit/8b78626daad10a8a40bb7ba4d3ca1b69c9756545). If possible, could we trigger a new PDF build? This is only a nice-to-have.
