### 2025-10-20 #307 @ashwinvis
@carschno There are some inconsistencies in AUTHORS and more importantly CITATION.cff file. I am fixing and will send a PR very soon, but you can decide whether to merge it now or after the acceptance.
