### 2025-10-20 1fd70cf
Add missing authors to Authors and CFF files
Namely Toby and Thor
