### 2025-10-20 #307 @ashwinvis
@carschno Here https://github.com/carpentries-lab/deep-learning-intro/pull/612
