### 2025-10-20 #307 @carschno
> [@carschno](https://github.com/carschno) There are some inconsistencies in AUTHORS and more importantly CITATION.cff file. I am fixing and will send a PR very soon, but you can decide whether to merge it now or after the acceptance.

Thanks for the PR, I have merged it.
Clarification for the reviewers: this latest PR has not made any changes to the paper, it just synchronizes the data in the lesson with the one specified in the paper.
