### 2025-10-23 5e4bb9a
Add new workshop entry for Helmholtz Zentrum
