### 2025-10-23 4e9b3f6
Update workshops.md
Co-authored-by: Carsten Schnober <carschno@gmail.com>
