### 2025-10-23 b090157
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 90e96d41b3b50a6852b58ff102a6385b7bea404b
Branch  : main
Author  : Gunjan Joshi <56921287+joshigunjan@users.noreply.github.com>
Time    : 2025-10-23 15:26:17 +0000
Message : Add new workshop entry for Helmholtz Zentrum (#613)

* Add new workshop entry for Helmholtz Zentrum
