### 2025-10-28 6d28c80
site deploy
Auto-generated via `{sandpaper}`
Source  : b090157521a01aa5dfdde01fafc6d567e0f9163c
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-10-23 15:26:57 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : 90e96d41b3b50a6852b58ff102a6385b7bea404b
Branch  : main
Author  : Gunjan Joshi <56921287+joshigunjan@users.noreply.github.com>
Time    : 2025-10-23 15:26:17 +0000
Message : Add new workshop entry for Helmholtz Zentrum (#613)

* Add new workshop entry for Helmholtz Zentrum
