### 2025-11-20 #307 @carschno
Hi @magsol ,
Just double-checking: now that all boxes are checked, is there anything left to do for the publication?
