### 2025-11-20 #307 @magsol
Hi @carschno! There is nothing left to do here; after adding the "recommend-accept" tag to this ticket, it's up to the JOSE EiCs to make a final pass and publish it. Sometimes this can take awhile, as our wonderful EiCs are busy folks. Sorry for the wait, but I appreciate your patience! I'll ping the internal Slack and see if we can get things moving :)
