### 2025-12-09 6599452
Use always-most-recent DOI
Co-authored-by: Carsten Schnober <c.schnober@esciencecenter.nl>
