+++ CITATION.cff
@@ -6,7 +6,7 @@ title: Introduction to deep learning
 message: Learn deep learning with Python
 identifiers:
   - type: doi
-    value: 10.5281/zenodo.8308392
+    value: 10.5281/zenodo.8308391
 type: software
 authors:
   - given-names: Sven
