### 2025-12-09 2d9fa4c
define DOI as identifier in CFF (#616)
* define DOI as identifier

* Use always-most-recent DOI

Co-authored-by: Carsten Schnober <c.schnober@esciencecenter.nl>

---------

Co-authored-by: Carsten Schnober <c.schnober@esciencecenter.nl>
