### 2025-12-09 07aeeed
Replace "Transfer Learning" PNG with Mermaid source. (#617)
