### 2025-12-09 1a0ad4e
markdown source builds
Auto-generated via `{sandpaper}`
Source  : 07aeeedb7edf28894469b8fcd7dde8ee7727c940
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-12-09 11:20:02 +0000
Message : Replace "Transfer Learning" PNG with Mermaid source. (#617)
