### 2025-12-09 3787cc0
site deploy
Auto-generated via `{sandpaper}`
Source  : 4a7f982fde811dc0cc3b002db99e311470c4382a
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2025-12-09 13:16:21 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : 085c1c0716f9f10f6b63cd08c4661bdfa717fcfc
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-12-09 13:15:32 +0000
Message : Add fine-tuning, transfer learning to glossary. (#619)
