### 2025-12-09 8ec7b4e
Fix badge as in CITATION.cff file
Also Zenodo seemed to have updated the links to their SVG file and their
recommeneded way for adding badges. I used the "all versions" doi as in
the CFF file.
