### 2025-12-09 7870ffe
Updated the introduction with instructions for using PyTorch as the Keras backend.
