### 2025-12-09 d3fb943
Deep learning networks as learning increasingly abstract representations of the input data (#622)
