### 2025-12-09 e88ac43
markdown source builds
Auto-generated via `{sandpaper}`
Source  : d3fb943db2e2f5c009637fcab0fb2c41bf7aabdf
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2025-12-09 15:20:55 +0000
Message : Deep learning networks as learning increasingly abstract representations of the input data (#622)
