### 2025-12-09 133225d
source commit: e88ac43456ff8c8ddeda5822fe022efd2084e664
