+++ paper.md
@@ -1,0 +1,252 @@
+---
+title: 'Introduction to deep learning: Carpentries-style hands-on lesson material for introducing researchers to deep learning'
+tags:
+  - Python
+  - deep learning
+  - machine learning
+  - Keras
+  - neural networks
+authors:
+  - name: Sven A. van der Burg
+    orcid: 0000-0003-1250-6968
+    affiliation: 1 # (Multiple affiliations must be quoted, like "1, 2")
+  - name: Pranav Chandramouli
+    orcid: 0000-0002-7896-2969
+    affiliation: 1
+  - name: Anne Fouilloux
+    orcid: 0000-0002-1784-2920
+    affiliation: 2
+  - name: Cunliang Geng
+    orcid: 0000-0002-1409-8358
+    affiliation: 1
+  - name: Toby Hodges
+    orcid: 0000-0003-1766-456X
+    affiliation: 8
+  - name: Florian Huber
+    orcid: 0000-0002-3535-9406
+    affiliation: "1, 3"
+  - name: Dafne van Kuppevelt
+    orcid: 0000-0002-2662-1994
+    affiliation: 1
+  - name: Ashwin Vishnu Mohanan
+    orcid: 0000-0002-2979-6327
+    affiliation: 9    
+  - name: Colin Sauze
+    orcid: 0000-0001-5368-9217
+    affiliation: 5
+  - name: Carsten Schnober
+    orcid: 0000-0001-9139-1577
+    affiliation: 1
+  - name: Djura Smits
+    orcid: 0000-0003-4096-0260
+    affiliation: 1
+  - name: Peter Steinbach
+    orcid: 0000-0002-4974-230X
+    affiliation: 4
+  - name: Berend Weel
+    orcid: 0000-0002-9693-9332
+    affiliation: 1
+  - name: Kjartan Thor Wikfeldt
+    orcid: 0000-0002-1655-3676
+    affiliation: 9
+  - name: Samantha Wittke
+    orcid: 0000-0002-9625-7235
+    affiliation: "6,7"
+    
+
+affiliations:
+ - name: Netherlands eScience Center, Amsterdam, The Netherlands
+   index: 1
+ - name: Simula Research Laboratory, Oslo, Norway
+   index: 2
+ - name: Düsseldorf University of Applied Sciences, Düsseldorf, Germany
+   index: 3
+ - name: Helmholtz-Zentrum Dresden-Rossendorf, Dresden, Germany
+   index: 4
+ - name: National Oceanography Centre, Liverpool, Great-Britain
+   index: 5
+ - name: CSC - IT center for Science, Espoo, Finland
+   index: 6
+ - name: Aalto University, Espoo, Finland
+   index: 7
+ - name: The Carpentries, USA
+   index: 8
+ - name: RISE Research Institutes of Sweden, Sweden
+   index: 9   
+date: 8 August 2023
+bibliography: paper.bib
+
+---
+
+# Summary
+This article describes a hands-on introduction to the first steps in deep learning, 
+intended for researchers who are familiar with (non-deep) machine learning.
+
+The use of deep learning has seen a sharp increase in popularity and applicability over the last decade. 
+While deep learning can be a useful tool for researchers from a wide range of domains, 
+taking the first steps in the world of deep learning can be somewhat intimidating. 
+This introduction aims to cover the fundamentals of deep learning in a practical and hands-on manner. By the end of the course, students will be able to train their first neural network and understand the subsequent steps needed to improve the model.
+
+The lesson starts by explaining the basic concepts of neural networks, 
+and then guides learners through the different steps of a deep learning workflow.  
+After following this lesson, 
+learners will be able to prepare data for deep learning, 
+implement a basic deep learning model in Python with Keras,
+and monitor and troubleshoot the training process.
[…]
