+++ 5-transfer-learning.md
@@ -1,0 +1,273 @@
+---
+title: "Transfer learning"
+teaching: 20
+exercises: 30
+---
+
+::: questions
+- How do I apply a pre-trained model to my data?
+:::
+
+::: objectives
+- Adapt a state-of-the-art pre-trained network to your own dataset
+:::
+
+## What is transfer learning?
+Instead of training a model from scratch, with transfer learning you make use of models that are trained on another machine learning task. The pre-trained network captures generic knowledge during pre-training and will only be 'fine-tuned' to the specifics of your dataset.
+
+An example: Let's say that you want to train a model to classify images of different dog breeds. You could make use of a pre-trained network that learned how to classify images of dogs and cats. The pre-trained network will not know anything about different dog breeds, but it will have captured some general knowledge of, on a high-level, what dogs look like, and on a low-level all the different features (eyes, ears, paws, fur) that make up an image of a dog. Further training this model on your dog breed dataset is a much easier task than training from scratch, because the model can use the general knowledge captured in the pre-trained network.
+
+```mermaid
+flowchart LR
+    accTitle: Transfer Learning
+    accDescr {The "Dogs & Cats" dataset is used to train a model.
+    With Transfer Learning, the smaller "Dog Breeds" dataset is used to add more specific data.
+    This results in the "Dog Breeds" model that makes use of both data sets.}
+    A[(Dogs & Cats Data)] --> |Train Model| B(Model Dogs & Cats)
+    C{Transfer Learning}
+    B --> C
+    D[(Dog Breeds Data)] --> C
+    C --> E(Dog Breeds Model)
+```
+
+In this episode we will learn how use Keras to adapt a state-of-the-art pre-trained model to the [Dollar Street Dataset](https://zenodo.org/records/10970014).
+
+
+## 1. Formulate / Outline the problem
+
+
+Just like in the previous episode, we use the Dollar Street 10 dataset. 
+
+We load the data in the same way as the previous episode:
+```python
+import pathlib
+import numpy as np
+
+DATA_FOLDER = pathlib.Path('data/dataset_dollarstreet/') # change to location where you stored the data
+train_images = np.load(DATA_FOLDER / 'train_images.npy')
+val_images = np.load(DATA_FOLDER / 'test_images.npy')
+train_labels = np.load(DATA_FOLDER / 'train_labels.npy')
+val_labels = np.load(DATA_FOLDER / 'test_labels.npy')
+```
+## 2. Identify inputs and outputs
+
+As discussed in the previous episode, the input are images of dimension 64 x 64 pixels with 3 colour channels each.
+The goal is to predict one out of 10 classes to which the image belongs.
+
+
+## 3. Prepare the data
+We prepare the data as before, scaling the values between 0 and 1.
+```python
+train_images = train_images / 255.0
+val_images = val_images / 255.0
+```
+
+## 4. Choose a pre-trained model or start building architecture from scratch
+Let's define our model input layer using the shape of our training images:
+```python
+# input tensor
+from tensorflow import keras
+keras.utils.set_random_seed(2)
+
+inputs = keras.Input(train_images.shape[1:])
+```
+
+Our images are 64 x 64 pixels, whereas the pre-trained model that we will use was 
+trained on images of 160 x 160 pixels.
+To adapt our data accordingly, we add an upscale layer that resizes the images to 160 x 160 pixels during training and prediction.
+
+```python
+# upscale layer
+import tensorflow as tf
+method = tf.image.ResizeMethod.BILINEAR
+upscale = keras.layers.Lambda(
+  lambda x: tf.image.resize_with_pad(x, 160, 160, method=method))(inputs)
+```
+
+From the `keras.applications` module we use the `DenseNet121` architecture. 
[…]

[…]
