+++ md5sum.txt
@@ -5,8 +5,8 @@
 "index.md" "8b5609014b8028029f48266bc751663e" "site/built/index.md" "2025-05-06"
 "links.md" "8184cf4149eafbf03ce8da8ff0778c14" "site/built/links.md" "2025-02-11"
 "workshops.md" "db285c697b5f062098913c08119130ff" "site/built/workshops.md" "2025-10-23"
-"episodes/1-introduction.md" "3602dbf5bab0ed4e21679de5f27d3f47" "site/built/1-introduction.md" "2025-12-09"
-"episodes/2-keras.md" "ab5ebd62fd3e6cc2ad69bd805d26f7f1" "site/built/2-keras.md" "2025-09-03"
+"episodes/1-introduction.md" "41c91d12c3e45b5bc5f339a0a4aae666" "site/built/1-introduction.md" "2025-12-09"
+"episodes/2-keras.md" "5583f21576ffb21ba9b37c7c5aab7add" "site/built/2-keras.md" "2025-12-09"
 "episodes/3-monitor-the-model.md" "8eda70a03e5225033ba05563b517ec88" "site/built/3-monitor-the-model.md" "2025-09-02"
 "episodes/4-advanced-layer-types.md" "eaca1f96ead3140467a4ba7f069284fa" "site/built/4-advanced-layer-types.md" "2025-09-03"
 "episodes/5-transfer-learning.md" "65ed8dff158123d8271100fd8d18fd3b" "site/built/5-transfer-learning.md" "2025-12-09"
@@ -17,6 +17,6 @@
 "instructors/schedule.md" "332b32d24f144b29a280176e6b5d015f" "site/built/schedule.md" "2025-03-10"
 "instructors/survey-templates.md" "ea5d46e7b54d335f79e57a7bc31d1c5c" "site/built/survey-templates.md" "2025-02-11"
 "learners/reference.md" "6a11d5269dc9d1d31d4016086580e838" "site/built/reference.md" "2025-12-09"
-"learners/setup.md" "2e741a6d76091da5c832ab010380f0e1" "site/built/setup.md" "2025-09-01"
+"learners/setup.md" "dbf2c7a18b7bc9ab5d404024b2e48930" "site/built/setup.md" "2025-12-09"
 "paper/paper.md" "2b05562e0f9d393818ad4e7ea2d5c152" "site/built/paper.md" "2025-10-20"
 "profiles/learner-profiles.md" "ef0f26dd0874387d80ed3fd468b99e23" "site/built/learner-profiles.md" "2025-02-11"
