### 2025-12-31 #307 @labarba
- [x] version v1.2.2
- [x] repo license CC-BY + MIT
- [x] Zenodo license shows CC-BY
- [ ] Zenodo title and authors
