### 2025-12-31 #307 @labarba
@carschno 👋  I noticed the author list doesn't quite match between the paper and the Zenodo archive. Can you check? (It appears it is just the ordering that is mangled.)
