### 2025-12-31 #307 @labarba
On the paper: I wonder if you could try making the table headings on Tables 1 and 2 lower case, to see if the layout can be improved. As it is, you have some overlapping headings and some margin bleed.
