### 2025-12-31 #307 @labarba
In the References you should add the author names for the listed courses, where available. For example, the author of the Coursera Deep Learning course is Andrew Yng; the author of the Fast.ai course is Jeremy Howard; the author of the freeCodeCamp course is Daniel Bourke.
