### 2026-01-02 c34d29f
Lowercase table headers.
