### 2026-01-02 1492e10
Fix CIFAR-10 citation.
