+++ paper/paper.bib
@@ -297,10 +297,13 @@ The original dataset was downloaded from https://www.kaggle.com/datasets/mlcommo
 	file = {Snapshot:/Users/svenvanderburg/Zotero/storage/QPJDIYXH/10970014.html:text/html},
 }
 
-@misc{noauthor_cifar-10_nodate,
-	title = {{CIFAR}-10 and {CIFAR}-100 datasets},
+
+@article{krizhevsky_learning_2009,
+	title = {Learning {Multiple} {Layers} of {Features} from {Tiny} {Images}},
 	url = {https://www.cs.toronto.edu/~kriz/cifar.html},
-	urldate = {2025-02-11},
-	file = {CIFAR-10 and CIFAR-100 datasets:/Users/svenvanderburg/Zotero/storage/CTXZX76B/cifar.html:text/html},
+	language = {en},
+	author = {Krizhevsky, Alex},
+	year = {2009},
 }
+
 
