### 2026-01-02 65a9ddc
Update Coursera Deep Learning reference.
