+++ paper/paper.bib
@@ -55,14 +55,17 @@
 	file = {Snapshot:/Users/svenvanderburg/Zotero/storage/EAUMLMBT/complete-deep-learning-course-with-python.html:text/html},
 }
 
-@misc{noauthor_coursera_nodate,
+
+@misc{ng_deep_nodate,
 	title = {Coursera - {Deep} {Learning}},
 	url = {https://www.coursera.org/specializations/deep-learning},
-	abstract = {Learn Deep Learning from deeplearning.ai. If you want to break into Artificial intelligence (AI), this Specialization will help you. Deep Learning is one of the most highly sought after skills in tech. We will help you become good at Deep Learning.},
+	abstract = {Offered by DeepLearning.AI. Become a Machine Learning expert. Master the fundamentals of deep learning and break into AI. Recently updated ... Enroll for free.},
 	language = {en},
-	urldate = {2023-09-01},
+	urldate = {2026-01-02},
 	journal = {Coursera},
+	author = {Ng, Andrew and Katanforoosh, Kian and Mourri, Younes},
 }
+
 
 @misc{noauthor_freecodecamporg_2022,
 	title = {{freeCodeCamp}.org - {Learn} {PyTorch} for {Deep} {Learning}},
