### 2026-01-02 436e8c2
Update Fast.ai reference.
