+++ paper/paper.bib
@@ -15,13 +15,13 @@
 	file = {The Carpentries Workbench:/Users/svenvanderburg/Zotero/storage/SSBZ6XPS/workbench.html:text/html},
 }
 
-@misc{noauthor_fastai_nodate,
-	title = {fast.ai - {Practical} {Deep} {Learning} for {Coders}},
-	url = {https://course.fast.ai/},
-	abstract = {A free course designed for people with some coding experience, who want to learn how to apply deep learning and machine learning to practical problems.},
-	language = {en},
-	urldate = {2023-09-01},
-	file = {Snapshot:/Users/svenvanderburg/Zotero/storage/CINJVAEW/course.fast.ai.html:text/html},
+@book{howard2020deep,
+	title={Deep Learning for Coders with Fastai and Pytorch: AI Applications Without a PhD},
+	author={Howard, J. and Gugger, S.},
+	isbn={9781492045526},
+	url={https://books.google.no/books?id=xd6LxgEACAAJ},
+	year={2020},
+	publisher={O'Reilly Media, Incorporated}
 }
 
 @misc{noauthor_udemy_nodate,
