+++ paper/paper.bib
@@ -66,14 +66,14 @@
 	author = {Ng, Andrew and Katanforoosh, Kian and Mourri, Younes},
 }
 
+@misc{freecodecamporg_pytorch_2022,
+	title = {{PyTorch} for {Deep} {Learning} \& {Machine} {Learning} – {Full} {Course}},
+	url = {https://www.youtube.com/watch?v=V_xro1bcAuA},
+	abstract = {Learn PyTorch for deep learning in this comprehensive course for beginners. PyTorch is a machine learning framework written in Python.
 
-@misc{noauthor_freecodecamporg_2022,
-	title = {{freeCodeCamp}.org - {Learn} {PyTorch} for {Deep} {Learning}},
-	url = {https://www.freecodecamp.org/news/learn-pytorch-for-deep-learning-in-day/},
-	abstract = {My comprehensive PyTorch course is now live on the freeCodeCamp.org YouTube channel.   * You can view the full 26 hour course here [https://youtu.be/V\_xro1bcAuA].  * Read the course materials online for free at learnpytorch.io    [https://learnpytorch.io/].  * See all of the course materials on GitHub},
-	language = {en},
-	urldate = {2023-09-01},
-	journal = {freeCodeCamp.org},
+✏️ Daniel Bourke developed this course. Check out his channel:    / @mrdbourke},
+	urldate = {2026-01-02},
+	author = {{freeCodeCamp.org}},
 	month = oct,
 	year = {2022},
 }
