### 2026-01-02 fdd0ef9
Update INRIA sciti learn Mooc reference.
