+++ paper/paper.bib
@@ -132,17 +132,17 @@
 	file = {Snapshot:/Users/svenvanderburg/Zotero/storage/227MJWAZ/CITATION.html:text/html},
 }
 
-@misc{noauthor_scikit-learn_2023,
-	title = {scikit-learn course},
-	copyright = {CC-BY-4.0},
-	url = {https://github.com/INRIA/scikit-learn-mooc},
+@misc{esteve_inriascikit-learn-mooc_2022,
+	title = {{INRIA}/scikit-learn-mooc: {Third} {MOOC} session},
+	shorttitle = {{INRIA}/scikit-learn-mooc},
+	url = {https://zenodo.org/records/7220307},
 	abstract = {Machine learning in Python with scikit-learn MOOC},
-	urldate = {2023-09-01},
-	publisher = {Inria},
-	month = sep,
-	year = {2023},
-	note = {original-date: 2020-03-09T14:53:36Z},
-	keywords = {machine-learning, mooc, python, scikit-learn},
+	urldate = {2026-01-02},
+	publisher = {Zenodo},
+	author = {Estève, Loïc and Lemaitre, Guillaume and Grisel, Olivier and Varoquaux, Gael and Amor, Arturo and Lilian and Rospars, Benoit and Schmitt, Thomas and Liu, Lucy and Kinoshita, Bruno P. and hackmd-deploy and ph4ge and Steinbach, Peter and Boucaud, Alexandre and Muite, Benson and Boisberranger, Jérémie du and Notter, Michael and Pierre and P, Shane and alagarrigue and Mohammadian, Mehrdad and parmentelat},
+	month = oct,
+	year = {2022},
+	doi = {10.5281/zenodo.7220307},
 }
 
 @software{Pollard_Introduction_to_artificial_2022,
