### 2026-01-02 4caac5d
Update Udemy Basics of Deep Learning reference.
