+++ paper/paper.bib
@@ -24,14 +24,14 @@
 	publisher={O'Reilly Media, Incorporated}
 }
 
-@misc{noauthor_udemy_nodate,
-	title = {Udemy - {Basics} of {Deep} {Learning}},
+@misc{mishra_free_nodate,
+	title = {Free {Deep} {Learning} {Tutorial} - {Basics} of {Deep} {Learning}},
 	url = {https://www.udemy.com/course/basics-of-deep-learning/},
 	abstract = {Fundamentals of Neural Network - Free Course},
 	language = {en-us},
-	urldate = {2023-09-01},
+	urldate = {2026-01-02},
 	journal = {Udemy},
-	file = {Snapshot:/Users/svenvanderburg/Zotero/storage/L57M3IZP/basics-of-deep-learning.html:text/html},
+	author = {Mishra, Sunil},
 }
 
 @misc{noauthor_udemy_nodate-1,
