### 2026-01-02 d5eebe3
Update Udemy Recurrent Neural Networks reference.
