+++ paper/paper.bib
@@ -76,17 +76,18 @@
 	author = {Ng, Andrew and Katanforoosh, Kian and Mourri, Younes},
 }
 
-@misc{freecodecamporg_pytorch_2022,
+@misc{bourke_pytorch_2022,
 	title = {{PyTorch} for {Deep} {Learning} \& {Machine} {Learning} – {Full} {Course}},
 	url = {https://www.youtube.com/watch?v=V_xro1bcAuA},
 	abstract = {Learn PyTorch for deep learning in this comprehensive course for beginners. PyTorch is a machine learning framework written in Python.
 
 ✏️ Daniel Bourke developed this course. Check out his channel:    / @mrdbourke},
 	urldate = {2026-01-02},
-	author = {{freeCodeCamp.org}},
+	author = {Bourke, Daniel},
 	month = oct,
 	year = {2022},
 }
+
 
 @misc{noauthor_csc-_nodate,
 	title = {{CSC}- {Practical} {Deep} {Learning}},
