+++ paper/paper.bib
@@ -304,7 +304,7 @@ Checkout this notebook to see how the subset was created.
 The original dataset was downloaded from https://www.kaggle.com/datasets/mlcommons/the-dollar-street-dataset. See https://mlcommons.org/datasets/dollar-street/ for more information.},
 	urldate = {2025-02-11},
 	publisher = {Zenodo},
-	author = {van der burg, Sven},
+	author = {van der Burg, Sven},
 	month = apr,
 	year = {2024},
 	keywords = {CC-BY, CIFAR-10, Deep learning, Image classification, Machine learning},
