### 2026-01-02 bccb21d
Fix Udemy course references.
