+++ paper/paper.bib
@@ -34,14 +34,14 @@
 	author = {Mishra, Sunil},
 }
 
-@misc{noauthor_udemy_nodate-1,
-	title = {Udemy - {Tensorflow} 2.0 {\textbar} {Recurrent} {Neural} {Networks}, {LSTMs}, {GRUs}},
+@misc{slim_free_nodate,
+	title = {Free {Deep} {Learning} {Tutorial} - {Tensorflow} 2.0 {\textbar} {Recurrent} {Neural} {Networks}, {LSTMs}, {GRUs}},
 	url = {https://www.udemy.com/course/tensorflow-20-recurrent-neural-networks-lstms-grus/},
 	abstract = {Sequence prediction course that covers topics such as: RNN, LSTM, GRU, NLP, Seq2Seq, Attention, Time series prediction - Free Course},
 	language = {en-us},
-	urldate = {2023-09-01},
+	urldate = {2026-01-02},
 	journal = {Udemy},
-	file = {Snapshot:/Users/svenvanderburg/Zotero/storage/GNQFE2IZ/tensorflow-20-recurrent-neural-networks-lstms-grus.html:text/html},
+	author = {Slim, Jad and Abdulkader, Amer},
 }
 
 @misc{noauthor_udemy_nodate-2,
@@ -54,17 +54,6 @@
 	journal = {Udemy},
 	file = {Snapshot:/Users/svenvanderburg/Zotero/storage/EAUMLMBT/complete-deep-learning-course-with-python.html:text/html},
 }
-
-@misc{slim_free_nodate,
-	title = {Free {Deep} {Learning} {Tutorial} - {Tensorflow} 2.0 {\textbar} {Recurrent} {Neural} {Networks}, {LSTMs}, {GRUs}},
-	url = {https://www.udemy.com/course/tensorflow-20-recurrent-neural-networks-lstms-grus/},
-	abstract = {Sequence prediction course that covers topics such as: RNN, LSTM, GRU, NLP, Seq2Seq, Attention, Time series prediction - Free Course},
-	language = {en-us},
-	urldate = {2026-01-02},
-	journal = {Udemy},
-	author = {Slim, Jad and Abdulkader, Amer},
-}
-
 
 @misc{ng_deep_nodate,
 	title = {Coursera - {Deep} {Learning}},
