### 2026-01-02 #307 @carschno
Good catch, thanks @labarba !

I have fixed both the table headings and the references in https://github.com/carpentries-lab/deep-learning-intro/pull/625 (merged). Some of the courses do not have authors, I have added the names for all that do.
Could you check again?

Regarding the author ordering, I do not have sufficient privileges to change them in the Zenodo entry. I am tagging @svenvanderburg, the original author. However, this is a long shot because he does not work here any longer, so I do not expect that we will be able to change that. Fortunately, it only affects the ordering, so I would suggest that we proceed with this slight inconsistency.
