### 2026-01-02 1fef2f8
site deploy
Auto-generated via `{sandpaper}`
Source  : 1c2e3ce5ae9cf5dac4f3cb55fed9d7eee7afda87
Branch  : md-outputs
Author  : GitHub Actions <actions@github.com>
Time    : 2026-01-02 11:28:33 +0000
Message : markdown source builds

Auto-generated via `{sandpaper}`
Source  : 65b6ae8d638cc2b68c40d359016964edd2a3d889
Branch  : main
Author  : Carsten Schnober <c.schnober@esciencecenter.nl>
Time    : 2026-01-02 11:27:45 +0000
Message : Merge pull request #625 from carpentries-lab/fix/tables

Paper Review
