### 2026-01-02 #307 @labarba
Congratulations @svenvanderburg, @carschno and coauthors — your JOSE paper is published! 🚀

Huge thanks to our Editor: @magsol and Reviewer: @manideepikaa — your contributions make this possible 🙏
