### 2026-01-02 #307 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
